refactor: split ai_tab.py (7927→4478 lines) into 8 focused mixin modules
Extract 8 cohesive groups of methods from the monolithic ai_tab.py into
separate mixin modules for maintainability — architecture-only change,
no behavior changes:

- ui/memory_mixin.py      (349 lines) — long-term memory & plugin hooks
- ui/context_mixin.py     (343 lines) — font scaling, token stats, provider/model mgmt
- ui/streaming_mixin.py   (339 lines) — streaming output & <think> tag parsing
- ui/plan_mixin.py        (322 lines) — plan mode UI logic
- ui/image_mixin.py       (251 lines) — multimodal image attachment & preview
- core/cache_mixin.py     (897 lines) — session cache save/restore/archive
- ui/history_mixin.py     (723 lines) — conversation history rendering & batch load
- core/update_mixin.py    (446 lines) — token optimization menu & auto-update

AITab inherits all 8 via Python multiple inheritance; PySide2 Qt signals
defined in plain mixin classes are picked up by QObject metaclass through MRO.
All 10 files pass ast.parse() syntax check; 0 duplicate methods detected.
