+++ houdini_agent/main.py
@@ -37,6 +37,14 @@ def _reload_modules():
         'houdini_agent.ui.chat_view',
         'houdini_agent.core.agent_runner',
         'houdini_agent.core.session_manager',
+        'houdini_agent.ui.memory_mixin',
+        'houdini_agent.ui.context_mixin',
+        'houdini_agent.ui.streaming_mixin',
+        'houdini_agent.ui.plan_mixin',
+        'houdini_agent.ui.image_mixin',
+        'houdini_agent.core.cache_mixin',
+        'houdini_agent.ui.history_mixin',
+        'houdini_agent.core.update_mixin',
         'houdini_agent.ui.ai_tab',
         'houdini_agent.core.main_window',
     ]
