+++ houdini_agent/ui/ai_tab.py
@@ -3,12 +3,20 @@
 Houdini Agent - AI Tab
 Agent loop, multi-turn tool calling, streaming UI
 
-模块拆分结构（逐步迁移中）:
-  ui/header.py          — HeaderMixin: 顶部设置栏构建
-  ui/input_area.py      — InputAreaMixin: 输入区域和模式切换
-  ui/chat_view.py       — ChatViewMixin: 对话显示和滚动逻辑
-  core/agent_runner.py  — AgentRunnerMixin: Agent 循环和工具调度
+模块拆分结构:
+  ui/header.py            — HeaderMixin: 顶部设置栏构建
+  ui/input_area.py        — InputAreaMixin: 输入区域和模式切换
+  ui/chat_view.py         — ChatViewMixin: 对话显示和滚动逻辑
+  core/agent_runner.py    — AgentRunnerMixin: Agent 循环和工具调度
   core/session_manager.py — SessionManagerMixin: 多会话管理和缓存
+  ui/memory_mixin.py      — MemoryMixin: 长期记忆 + 插件 Hook 系统
+  ui/context_mixin.py     — ContextMixin: 字体缩放、上下文统计、模型管理
+  ui/streaming_mixin.py   — StreamingMixin: 流式内容 + <think> 解析
+  ui/plan_mixin.py        — PlanMixin: 计划模式 UI 逻辑
+  ui/image_mixin.py       — ImageMixin: 多模态图片处理
+  core/cache_mixin.py     — CacheMixin: 会话缓存保存/恢复/存档
+  ui/history_mixin.py     — HistoryMixin: 对话历史渲染
+  core/update_mixin.py    — UpdateMixin: Token 优化菜单 + 自动更新
 """
 
 import json
@@ -66,14 +74,16 @@ from .input_area import InputAreaMixin
 from .chat_view import ChatViewMixin
 from ..core.agent_runner import AgentRunnerMixin
 from ..core.session_manager import SessionManagerMixin
+from .memory_mixin import MemoryMixin
+from .context_mixin import ContextMixin
+from .streaming_mixin import StreamingMixin
+from .plan_mixin import PlanMixin
+from .image_mixin import ImageMixin
+from ..core.cache_mixin import CacheMixin
+from .history_mixin import HistoryMixin
+from ..core.update_mixin import UpdateMixin
 
-# ★ 大脑启发式长期记忆系统
-from ..utils.memory_store import get_memory_store
-from ..utils.reward_engine import get_reward_engine
-from ..utils.reflection import get_reflection_module
-from ..utils.growth_tracker import get_growth_tracker, TaskMetric
-
-# ★ Plan 模式
+# ★ Plan 模式常量（agent_runner.py 中也需要引用）
 from ..utils.plan_manager import get_plan_manager, PLAN_TOOL_CREATE, PLAN_TOOL_UPDATE_STEP, PLAN_TOOL_ASK_QUESTION
 
 
@@ -83,6 +93,14 @@ class AITab(
     ChatViewMixin,
     AgentRunnerMixin,
     SessionManagerMixin,
+    MemoryMixin,
+    ContextMixin,
+    StreamingMixin,
+    PlanMixin,
+    ImageMixin,
+    CacheMixin,
+    HistoryMixin,
+    UpdateMixin,
     QtWidgets.QWidget,
 ):
     """AI 助手 - 极简侧边栏风格（Mixin 架构）"""
@@ -351,330 +369,6 @@ class AITab(
     # ==========================================================
     # ★ 大脑启发式长期记忆系统
     # ==========================================================
-
-    def _init_memory_system(self):
-        """初始化长期记忆系统（后台线程，不阻塞 UI）
-
-        注意：初始化始终进行（成本低、允许用户随时打开开关），
-        但实际的注入/反思/睡眠只在 self._memory_enabled 为 True 时触发。
-        """
-        def _init():
-            try:
-                self._memory_store = get_memory_store()
-                self._reward_engine = get_reward_engine()
-                self._reflection_module = get_reflection_module()
-                self._growth_tracker = get_growth_tracker()
-                self._memory_initialized = True
-                print(f"[Memory] 长期记忆系统已初始化 (enabled={self._memory_enabled}): "
-                      f"{self._memory_store.get_stats()}")
-            except Exception as e:
-                print(f"[Memory] 初始化失败 (非致命): {e}")
-                self._memory_initialized = False
-
-        thread = threading.Thread(target=_init, daemon=True)
-        thread.start()
-
-    # ---------- 全局开关：记忆系统启用/禁用 ----------
-
-    @staticmethod
-    def _load_memory_enabled_pref() -> bool:
-        """从 QSettings 加载记忆开关（默认 False）。"""
-        settings = QSettings("HoudiniAI", "Assistant")
-        val = settings.value("memory_enabled", False)
-        if isinstance(val, str):
-            return val.lower() == 'true'
-        return bool(val)
-
-    def _save_memory_enabled_pref(self, enabled: bool):
-        settings = QSettings("HoudiniAI", "Assistant")
-        settings.setValue("memory_enabled", bool(enabled))
-
-    def _is_memory_active(self) -> bool:
-        """记忆相关钩子的统一短路条件。
-
-        True 时才应注入 L0 核心记忆、激活分层检索、反思、睡眠以及
-        暴露 search_memory 工具；False 时完全关闭。
-        """
-        return bool(self._memory_enabled and self._memory_initialized and self._memory_store)
-
-    def set_memory_enabled(self, enabled: bool):
-        """切换记忆系统全局开关并持久化。"""
-        enabled = bool(enabled)
-        if enabled == self._memory_enabled:
-            return
-        self._memory_enabled = enabled
-        self._save_memory_enabled_pref(enabled)
-        # 状态栏提示
-        key = 'memory.toggle.enabled' if enabled else 'memory.toggle.disabled'
-        try:
-            self._addStatus.emit(tr(key))
-        except Exception:
-            pass
-
-    # ==========================================================
-    # ★ 插件系统 (Hook / Plugin System)
-    # ==========================================================
-
-    def _init_plugin_system(self):
-        """初始化插件系统：加载插件、设置 UI Bridge、挂载按钮"""
-        try:
-            from ..utils.hooks import get_hook_manager, PluginUIBridge, load_all_plugins
-
-            manager = get_hook_manager()
-
-            # 创建 UI Bridge 并关联到 HookManager
-            bridge = PluginUIBridge()
-            # 设置按钮容器引用
-            if hasattr(self, '_plugin_button_container'):
-                bridge.set_button_container(self._plugin_button_container)
-            # 设置聊天区域布局（供 insert_chat_card 使用）
-            if hasattr(self, 'chat_layout') and self.chat_layout:
-                bridge.set_chat_layout(self.chat_layout)
-            bridge.set_ai_tab(self)
-            manager.set_ui_bridge(bridge)
-
-            # 加载所有插件
-            load_all_plugins()
-
-            # 挂载插件按钮
-            bridge.mount_buttons()
-
-            print("[Hook] 插件系统初始化完成")
-        except Exception as e:
-            print(f"[Hook] 插件系统初始化失败 (非致命): {e}")
-
-    def _fire_session_hook(self, event: str, session_id: str):
-        """触发会话相关的 Hook 事件"""
-        try:
-            from ..utils.hooks import get_hook_manager
-            get_hook_manager().fire(event, session_id=session_id)
-        except Exception:
-            pass
-
-    def _activate_long_term_memory(self, user_message: str, scene_context: dict = None) -> str:
-        """动态记忆激活 — 分层 chunk 检索
-
-        6 层抽象层级体系：
-        - L0 (核心身份): 已在 sys_prompt 中加载，此处跳过
-        - L1 (核心偏好): embedding 检索, top_k=3, threshold=0.15
-        - L2 (经验规则): embedding 检索, top_k=3, threshold=0.25
-        - L3 (工作流模式): embedding 检索, top_k=2, threshold=0.35
-        - L4-L5: 不自动注入，仅通过 search_memory 工具检索
-
-        每层独立取 TopK chunk，互不挤占。
-        每条 chunk 附带置信度标注，明确标注"仅供参考"。
-
-        ★ 注意: fallback embedding (n-gram hash) 的 cosine similarity 值域约 0~0.4，
-        远低于 sentence-transformers 的 0~1.0。threshold 会在 search_by_level 内部
-        自动缩放以适配不同后端。Episodic / Procedural 的 score 阈值也需同样处理。
-        """
-        if not self._is_memory_active():
-            return ""
-
-        try:
-            store = self._memory_store
-
-            # 构建查询（用户消息 + 场景关键词）
-            query = user_message
-            if scene_context:
-                selected_types = scene_context.get('selected_types', [])
-                if selected_types:
-                    query += ' ' + ' '.join(selected_types)
-
-            # ★ fallback 模式下 cosine similarity 值域很低，缩放 score 阈值
-            _is_semantic = store.embedder.is_semantic
-            _ep_threshold = 0.3 if _is_semantic else 0.05
-            _proc_threshold = 0.25 if _is_semantic else 0.04
-
-            parts = []
-
-            # ── L1: 核心偏好 (top_k=3, threshold=0.15) ──
-            l1_results = store.search_by_level(query, level=1, top_k=3, threshold=0.15)
-            for rec, score in l1_results:
-                parts.append(f"[L1 Preference] (conf={rec.confidence:.2f}) {rec.rule[:120]}")
-                store.increment_semantic_activation(rec.id)
-
-            # ── L2: 经验规则 (top_k=3, threshold=0.25) ──
-            l2_results = store.search_by_level(query, level=2, top_k=3, threshold=0.25)
-            for rec, score in l2_results:
-                parts.append(f"[L2 Rule] (conf={rec.confidence:.2f}) {rec.rule[:120]}")
-                store.increment_semantic_activation(rec.id)
-
-            # ── L3: 工作流模式 (top_k=2, threshold=0.35) ──
-            l3_results = store.search_by_level(query, level=3, top_k=2, threshold=0.35)
-            for rec, score in l3_results:
-                parts.append(f"[L3 Workflow] (conf={rec.confidence:.2f}) {rec.rule[:120]}")
-                store.increment_semantic_activation(rec.id)
-
-            # ── Episodic: 相关经历 (top_k=2) ──
-            episodes = store.search_episodic(query, top_k=2, min_importance=0.3)
-            for ep, score in episodes:
-                if score > _ep_threshold:
-                    status = "✅" if ep.success else "❌"
-                    parts.append(
-                        f"[Past Experience] {status} {ep.task_description[:80]} "
-                        f"→ {ep.result_summary[:60]}"
-                    )
-                    try:
-                        new_imp = min(5.0, ep.importance * 1.05)
-                        store.update_episodic_importance(ep.id, new_imp)
-                    except Exception:
-                        pass
-
-            # ── Procedural: 适用策略 (top_k=2) ──
-            strategies = store.search_procedural(query, top_k=2)
-            for strat, score in strategies:
-                if score > _proc_threshold:
-                    parts.append(f"[Strategy] {strat.description[:80]}")
-
-            if not parts:
-                return ""
-
-            header = "[Long-Term Memory — 历史经验仅供参考，请结合当前上下文判断]"
-            result = header + "\n" + "\n".join(parts)
-            return result
-
-        except Exception as e:
-            print(f"[Memory] 记忆激活失败: {e}")
-            return ""
-
-    @staticmethod
-    def _collect_recent_rounds(history: list, n_rounds: int) -> list:
-        """从对话历史中收集最近 N 轮（以 user 消息为分界）的消息
-
-        Args:
-            history: 完整对话历史
-            n_rounds: 要收集的轮数
-
-        Returns:
-            最近 N 轮的消息副本列表
-        """
-        if not history:
-            return []
-
-        # 按 user 消息划分轮次
-        rounds = []
-        current_round = []
-        for m in history:
-            if m.get('role') == 'user' and current_round:
-                rounds.append(current_round)
-                current_round = []
-            current_round.append(m)
-        if current_round:
-            rounds.append(current_round)
-
-        # 取最近 n_rounds 轮
-        recent = rounds[-n_rounds:] if len(rounds) >= n_rounds else rounds
-        # 展平为消息列表（深拷贝避免修改原始数据）
-        import copy
-        return [copy.copy(m) for rnd in recent for m in rnd]
-
-    def _reflect_after_task(self, result: dict, agent_params: dict):
-        """任务完成后的反思钩子 — 在后台线程执行
-
-        从 agent result 中提取信号，创建 episodic 记忆，
-        计算 reward，触发规则/LLM 反思。
-        """
-        if not self._is_memory_active() or not self._reflection_module:
-            return
-
-        try:
-            # 提取任务信息
-            tool_calls_history = result.get('tool_calls_history', [])
-            final_content = result.get('final_content', '') or result.get('content', '')
-            new_messages = result.get('new_messages', [])
-
-            # 构建工具调用序列
-            tool_calls = []
-            error_count = 0
-            retry_count = 0
-            for tc in tool_calls_history:
-                tc_result = tc.get('result', {})
-                success = bool(tc_result.get('success', True))
-                has_error = bool(tc_result.get('error', ''))
-                tool_calls.append({
-                    "name": tc.get('tool_name', ''),
-                    "success": success and not has_error,
-                    "error": tc_result.get('error', ''),
-                })
-                if has_error or not success:
-                    error_count += 1
-
-            # 检测重试（连续相同工具调用）
-            for i in range(1, len(tool_calls)):
-                if (tool_calls[i]["name"] == tool_calls[i-1]["name"]
-                        and not tool_calls[i-1]["success"]):
-                    retry_count += 1
-
-            # 提取用户请求
-            history = self._agent_history if self._agent_history is not None else self._conversation_history
-            task_description = ""
-            for msg in reversed(history):
-                if msg.get('role') == 'user':
-                    content = msg.get('content', '')
-                    if isinstance(content, list):
-                        task_description = ' '.join(
-                            p.get('text', '') for p in content if p.get('type') == 'text'
-                        )
-                    else:
-                        task_description = content
-                    task_description = task_description[:200]
-                    break
-
-            # 判断成功 / 失败
-            success = result.get('ok', True) and error_count < len(tool_calls) * 0.5
-
-            # 结果摘要
-            result_summary = ""
-            if final_content:
-                # 去除 think 标签
-                import re as _re
-                clean = _re.sub(r'<think>[\s\S]*?</think>', '', final_content).strip()
-                result_summary = clean[:150]
-
-            session_id = self._agent_session_id or self._session_id
-
-            # 执行反思
-            reflect_result = self._reflection_module.reflect_on_task(
-                session_id=session_id,
-                task_description=task_description,
-                result_summary=result_summary,
-                success=success,
-                error_count=error_count,
-                retry_count=retry_count,
-                tool_calls=tool_calls,
-                ai_client=self.client,
-                model=agent_params.get('model', 'deepseek-v4-flash'),
-                provider=agent_params.get('provider', 'deepseek'),
-            )
-
-            # 更新 Growth Tracker
-            if self._growth_tracker:
-                metric = TaskMetric(
-                    success=success,
-                    error_count=error_count,
-                    retry_count=retry_count,
-                    tool_call_count=len(tool_calls),
-                    reward=reflect_result.get('reward', 0.0),
-                    tags=reflect_result.get('tags', []),
-                )
-                self._growth_tracker.record_task(metric)
-
-                # 如果 LLM 反思返回了技能置信度更新
-                if reflect_result.get('deep_reflected') and 'skill_confidence' in reflect_result:
-                    self._growth_tracker.update_skill_confidence_batch(
-                        reflect_result.get('skill_confidence', {})
-                    )
-
-            if reflect_result.get('reward', 0) > 0:
-                print(f"[Memory] 反思完成: reward={reflect_result['reward']:.2f}, "
-                      f"tags={reflect_result.get('tags', [])}, "
-                      f"deep_reflected={reflect_result.get('deep_reflected', False)}")
-
-        except Exception as e:
-            import traceback
-            print(f"[Memory] 反思钩子异常: {e}")
-            traceback.print_exc()
 
     def _get_personality_injection(self) -> str:
         """获取个性注入文本（附加到 system prompt 末尾）"""
@@ -1041,326 +735,6 @@ SideFX Labs Node Usage Rules (MUST follow strictly):
 
     # ===== 字号缩放 =====
 
-    def _apply_font_scale(self):
-        """重新渲染 QSS 并应用到界面"""
-        self.setStyleSheet(self._theme.render())
-        self._theme.save_preference()
-
-    def _zoom_in(self):
-        self._theme.zoom_in()
-        self._apply_font_scale()
-
-    def _zoom_out(self):
-        self._theme.zoom_out()
-        self._apply_font_scale()
-
-    def _zoom_reset(self):
-        self._theme.zoom_reset()
-        self._apply_font_scale()
-
-    def _on_font_settings(self):
-        """打开字号设置面板"""
-        dlg = FontSettingsDialog(current_scale=self._theme.scale, parent=self)
-        dlg.scaleChanged.connect(self._on_font_scale_preview)
-        dlg.exec_()
-        # 对话框关闭后保存最终结果
-        self._theme.set_scale(dlg.scale)
-        self._apply_font_scale()
-
-    def _on_font_scale_preview(self, scale: float):
-        """实时预览字号缩放"""
-        self._theme.set_scale(scale)
-        self.setStyleSheet(self._theme.render())
-
-    # ===== 上下文统计 =====
-    
-    def _estimate_tokens(self, text: str) -> int:
-        """估算文本的 token 数量（粗略估算）
-        
-        中文约 1.5 字符/token，英文约 4 字符/token
-        这里使用简单的混合估算
-        """
-        if not text:
-            return 0
-        
-        # 统计中文字符
-        chinese_chars = sum(1 for c in text if '\u4e00' <= c <= '\u9fff')
-        other_chars = len(text) - chinese_chars
-        
-        # 中文约 1.5 字符/token，其他约 4 字符/token
-        tokens = chinese_chars / 1.5 + other_chars / 4
-        return int(tokens)
-    
-    def _calculate_context_tokens(self) -> int:
-        """计算当前上下文的总 token 数（含工具定义）"""
-        # 缓存工具定义 token 数（只算一次，因为工具定义不变）
-        if not hasattr(self, '_tools_token_cache'):
-            import json as _json
-            from houdini_agent.utils.ai_client import HOUDINI_TOOLS
-            tools_json = _json.dumps(HOUDINI_TOOLS, ensure_ascii=False)
-            self._tools_token_cache = self.token_optimizer.estimate_tokens(tools_json)
-        
-        total = self._tools_token_cache
-        
-        # 系统提示词
-        total += self.token_optimizer.estimate_tokens(self._system_prompt)
-        
-        # 上下文摘要
-        if self._context_summary:
-            total += self.token_optimizer.estimate_tokens(self._context_summary)
-        
-        # 对话历史
-        total += self.token_optimizer.calculate_message_tokens(self._conversation_history)
-        
-        return total
-    
-    def _save_model_preference(self):
-        """保存模型选择偏好"""
-        settings = QSettings("HoudiniAI", "Assistant")
-        provider = self._current_provider()
-        model = self.model_combo.currentText()
-        settings.setValue("last_provider", provider)
-        settings.setValue("last_model", model)
-        settings.setValue("use_think", self.think_check.isChecked())
-    
-    def _load_model_preference(self, restore_provider: bool = False):
-        """加载模型选择偏好
-        
-        Args:
-            restore_provider: 是否同时恢复提供商选择（仅在初始化时为 True）
-        """
-        settings = QSettings("HoudiniAI", "Assistant")
-        last_provider = settings.value("last_provider", "")
-        last_model = settings.value("last_model", "")
-        
-        # 恢复 Think 开关
-        use_think = settings.value("use_think", True)
-        # QSettings 可能返回字符串 "true"/"false"
-        if isinstance(use_think, str):
-            use_think = use_think.lower() == 'true'
-        self.think_check.setChecked(bool(use_think))
-        
-        if not last_provider:
-            return
-        
-        # 恢复提供商（仅在启动时调用一次）
-        if restore_provider and last_provider != self._current_provider():
-            for i in range(self.provider_combo.count()):
-                if self.provider_combo.itemData(i) == last_provider:
-                    # 暂时阻断信号，避免触发 _on_provider_changed 递归
-                    self.provider_combo.blockSignals(True)
-                    self.provider_combo.setCurrentIndex(i)
-                    self.provider_combo.blockSignals(False)
-                    # 手动刷新模型列表和状态
-                    self._refresh_models(last_provider)
-                    self._update_key_status()
-                    break
-        
-        # 恢复模型
-        current_provider = self._current_provider()
-        if last_provider == current_provider and last_model:
-            available_models = [self.model_combo.itemText(i) for i in range(self.model_combo.count())]
-            if last_model in available_models:
-                index = self.model_combo.findText(last_model)
-                if index >= 0:
-                    self.model_combo.setCurrentIndex(index)
-    
-    def _get_current_context_limit(self) -> int:
-        """获取当前模型的上下文限制"""
-        model = self.model_combo.currentText()
-        return self._model_context_limits.get(model, 64000)
-    
-    def _update_context_stats(self):
-        """更新上下文统计显示（包含优化状态）"""
-        used = self._calculate_context_tokens()
-        limit = self._get_current_context_limit()
-        
-        # 格式化显示
-        if used >= 1000:
-            used_str = f"{used / 1000:.1f}K"
-        else:
-            used_str = str(used)
-        
-        limit_str = f"{limit // 1000}K"
-        
-        # 计算百分比
-        percent = (used / limit) * 100 if limit > 0 else 0
-        
-        # 优化状态指示
-        optimize_indicator = ""
-        if self._auto_optimize:
-            should_compress, _ = self.token_optimizer.should_compress(used, limit)
-            if should_compress:
-                optimize_indicator = " *"  # 需要优化
-            else:
-                optimize_indicator = ""  # 已优化/正常
-        
-        # 根据使用比例设置颜色
-        if percent < 50:
-            color = CursorTheme.TEXT_MUTED
-        elif percent < 80:
-            color = CursorTheme.ACCENT_ORANGE
-        else:
-            color = CursorTheme.ACCENT_RED
-        
-        self.context_label.setText(f"{percent:.1f}% {used_str}/{limit_str}{optimize_indicator}")
-        # 动态状态 → QSS 选择器 QLabel#contextLabel[state="..."]
-        if percent < 50:
-            ctx_state = ""
-        elif percent < 80:
-            ctx_state = "warning"
-        else:
-            ctx_state = "critical"
-        self.context_label.setProperty("state", ctx_state)
-        self.context_label.style().unpolish(self.context_label)
-        self.context_label.style().polish(self.context_label)
-        
-        # 更新优化按钮状态（如果超过阈值，高亮显示）
-        opt_state = "warning" if percent >= 80 else ""
-        self.btn_optimize.setProperty("state", opt_state)
-        self.btn_optimize.style().unpolish(self.btn_optimize)
-        self.btn_optimize.style().polish(self.btn_optimize)
-
-    def _update_token_stats_display(self):
-        """更新 Token 统计按钮显示（对齐 Cursor：显示费用）"""
-        total = self._token_stats['total_tokens']
-        cost = self._token_stats.get('estimated_cost', 0.0)
-        
-        # 格式化 token 显示
-        if total >= 1000000:
-            tok_display = f"{total / 1000000:.1f}M"
-        elif total >= 1000:
-            tok_display = f"{total / 1000:.1f}K"
-        else:
-            tok_display = str(total)
-        
-        # 格式化费用显示（Cursor 风格：$0.12）
-        if cost >= 1.0:
-            cost_display = f"${cost:.2f}"
-        elif cost >= 0.01:
-            cost_display = f"${cost:.2f}"
-        elif cost > 0:
-            cost_display = f"${cost:.4f}"
-        else:
-            cost_display = ""
-        
-        # 按钮文本：token数 | $费用
-        if cost_display:
-            self.token_stats_btn.setText(f"{tok_display} | {cost_display}")
-        else:
-            self.token_stats_btn.setText(tok_display)
-        
-        # 计算 cache 命中率
-        cache_read = self._token_stats['cache_read']
-        cache_write = self._token_stats['cache_write']
-        cache_total = cache_read + cache_write
-        hit_rate_display = f"{(cache_read / cache_total * 100):.1f}%" if cache_total > 0 else "N/A"
-        
-        reasoning = self._token_stats.get('reasoning_tokens', 0)
-        reasoning_line = tr('token.reasoning_line', reasoning) if reasoning > 0 else ""
-        
-        self.token_stats_btn.setToolTip(
-            tr('token.summary',
-               self._token_stats['requests'],
-               self._token_stats['input_tokens'],
-               self._token_stats['output_tokens'],
-               reasoning_line,
-               cache_read, cache_write, hit_rate_display,
-               total, cost_display or '$0.00')
-        )
-    
-    def _show_token_stats_dialog(self):
-        """显示详细 Token 统计对话框（对齐 Cursor：使用 TokenAnalyticsPanel）"""
-        from houdini_agent.ui.cursor_widgets import TokenAnalyticsPanel
-        records = getattr(self, '_call_records', []) or []
-        dialog = TokenAnalyticsPanel(records, self._token_stats, parent=self)
-        dialog.exec_()
-        if dialog.should_reset_stats:
-            self._reset_token_stats()
-    
-    def _reset_token_stats(self):
-        """重置 Token 统计"""
-        self._token_stats = {
-            'input_tokens': 0,
-            'output_tokens': 0,
-            'reasoning_tokens': 0,
-            'cache_read': 0,
-            'cache_write': 0,
-            'total_tokens': 0,
-            'requests': 0,
-            'estimated_cost': 0.0,
-        }
-        self._call_records = []
-        self._update_token_stats_display()
-        
-        # 显示提示
-        if self._current_response:
-            self._current_response.add_status(tr('status.stats_reset'))
-
-    # ===== UI 辅助 =====
-    
-    def _current_provider(self) -> str:
-        return self.provider_combo.currentData() or 'deepseek'
-
-    def _refresh_models(self, provider: str):
-        self.model_combo.clear()
-
-        if provider == 'ollama':
-            # 后台拉取 Ollama 模型列表，避免主线程阻塞
-            import threading
-            self.model_combo.addItem("检测中...")
-            self.model_combo.setEnabled(False)
-            def _fetch():
-                try:
-                    models = self.client.get_ollama_models()
-                except Exception:
-                    models = []
-                QtCore.QTimer.singleShot(0, lambda: self._on_ollama_models_ready(models))
-            threading.Thread(target=_fetch, daemon=True).start()
-            return
-
-        # 使用预设的模型列表
-        self.model_combo.addItems(self._model_map.get(provider, []))
-
-    def _on_ollama_models_ready(self, models: list):
-        """Ollama 模型列表后台加载完成回调（主线程）"""
-        self.model_combo.setEnabled(True)
-        self.model_combo.clear()
-        if models:
-            self.model_combo.addItems(models)
-        else:
-            self.model_combo.addItems(self._model_map.get('ollama', []))
-        self._load_model_preference()
-
-    def _update_key_status(self):
-        provider = self._current_provider()
-        
-        if provider == 'ollama':
-            # 测试 Ollama 连接
-            result = self.client.test_connection('ollama')
-            if result.get('ok'):
-                self.key_status.setText("Local")
-                self.key_status.setProperty("state", "ok")
-            else:
-                self.key_status.setText("Offline")
-                self.key_status.setProperty("state", "error")
-        elif self.client.has_api_key(provider):
-            masked = self.client.get_masked_key(provider)
-            self.key_status.setText(masked)
-            self.key_status.setProperty("state", "ok")
-        else:
-            self.key_status.setText("No Key")
-            self.key_status.setProperty("state", "warning")
-        self.key_status.style().unpolish(self.key_status)
-        self.key_status.style().polish(self.key_status)
-
-    def _on_provider_changed(self):
-        provider = self._current_provider()
-        self._refresh_models(provider)
-        self._load_model_preference()  # 切换提供商时也尝试加载上次使用的模型
-        self._update_key_status()
-        self._on_provider_changed_custom_visibility()  # Custom ⚙ 按钮可见性
-
     def _set_running(self, running: bool):
         self._is_running = running
         
@@ -1512,323 +886,6 @@ SideFX Labs Node Usage Rules (MUST follow strictly):
 
     # ===== 信号处理 =====
     
-    def _on_append_content(self, text: str):
-        """处理内容追加（主线程槽函数）
-        
-        注意：内容已经在 _on_content_with_limit → _drain_tag_buffer → 
-        _emit_normal_content 中经过了 <think> 标签过滤和伪造检测。
-        这里只负责将文本交给 UI 控件显示，不做额外过滤。
-        """
-        resp = self._agent_response or self._current_response
-        if not text or not resp:
-            return
-        # ★ 修复：不丢弃包含换行符的 chunk
-        # 纯换行符（\n\n）是 Markdown 段落分隔的关键信号，
-        # 丢弃它们会导致多段内容粘连在一起
-        if not text.strip() and '\n' not in text:
-            return
-        try:
-            # ★ 内容开始流入 → 隐藏 "Generating..." 状态（如果正在显示）
-            if hasattr(self, 'thinking_bar') and getattr(self.thinking_bar, '_mode', None) == 'generating':
-                self.thinking_bar.stop()
-            resp.append_content(text)
-            self._scroll_agent_to_bottom(force=False)
-        except RuntimeError:
-            pass  # widget 已被 clear 销毁
-
-    def _on_content_with_limit(self, text: str):
-        """处理内容追加，解析 <think> 标签，分离思考和正式内容"""
-        if not text:
-            return
-
-        # 初始化输出缓冲
-        if not hasattr(self, '_output_buffer'):
-            self._output_buffer = ""
-            self._last_flush_time = time.time()
-            self._adaptive_buf_size = 80
-            self._adaptive_interval = 0.15
-            self._last_render_duration = 0.0
-            self._flush_count = 0
-            self._is_first_content_chunk = True
-
-        # 追加到标签解析缓冲区
-        self._tag_parse_buf += text
-        self._drain_tag_buffer()
-
-    # ------------------------------------------------------------------
-    # <think> 标签流式解析
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _partial_tag_at_end(text: str, tag: str) -> int:
-        """检测 text 末尾是否有 tag 的不完整前缀，返回匹配长度 (0 = 无)"""
-        for i in range(min(len(tag) - 1, len(text)), 0, -1):
-            if tag[:i] == text[-i:]:
-                return i
-        return 0
-
-    def _drain_tag_buffer(self):
-        """处理 _tag_parse_buf，将内容分发到正式输出或思考面板"""
-        buf = self._tag_parse_buf
-        while buf:
-            if not self._in_think_block:
-                # ── 正常模式：寻找 <think> ──
-                pos = buf.find('<think>')
-                if pos >= 0:
-                    if pos > 0:
-                        self._emit_normal_content(buf[:pos])
-                    buf = buf[pos + 7:]          # 跳过 <think>
-                    self._in_think_block = True
-                    # ★ Think 开关打开时才显示思考面板；关闭时静默丢弃 <think> 内容
-                    if self._think_enabled:
-                        self._thinking_needs_finalize = True  # 进入思考，标记需要 finalize
-                        # 如果思考已 finalize，恢复为活跃状态并重启计时
-                        self._resume_thinking()
-                    continue
-                # 检查末尾是否有不完整的 <think>
-                hold = self._partial_tag_at_end(buf, '<think>')
-                if hold:
-                    self._emit_normal_content(buf[:-hold])
-                    self._tag_parse_buf = buf[-hold:]
-                    return
-                # 全部是正常内容
-                self._emit_normal_content(buf)
-                self._tag_parse_buf = ""
-                return
-            else:
-                # ── 思考模式：寻找 </think> ──
-                pos = buf.find('</think>')
-                if pos >= 0:
-                    if self._think_enabled and pos > 0:
-                        self._addThinking.emit(buf[:pos])
-                    buf = buf[pos + 8:]          # 跳过 </think>
-                    self._in_think_block = False
-                    # 思考结束：立即 finalize 思考区块并停止计时器
-                    if self._think_enabled:
-                        self._finalize_thinking()
-                    continue
-                # 检查末尾是否有不完整的 </think>
-                hold = self._partial_tag_at_end(buf, '</think>')
-                if hold:
-                    if self._think_enabled:
-                        safe = buf[:-hold]
-                        if safe:
-                            self._addThinking.emit(safe)
-                    self._tag_parse_buf = buf[-hold:]
-                    return
-                # 全部是思考内容
-                if self._think_enabled:
-                    self._addThinking.emit(buf)
-                # ★ Think 关闭时：静默丢弃 <think> 块内的内容
-                self._tag_parse_buf = ""
-                return
-        self._tag_parse_buf = ""
-
-    def _finalize_thinking(self):
-        """思考阶段结束（线程安全：自动分派到主线程）"""
-        self._finalizeThinkingSignal.emit()
-
-    def _resume_thinking(self):
-        """新一轮 <think> 开始（线程安全：自动分派到主线程）"""
-        self._resumeThinkingSignal.emit()
-
-    @QtCore.Slot()
-    def _finalize_thinking_main_thread(self):
-        """[主线程] 实际执行 finalize 思考区块并停止计时器"""
-        try:
-            resp = self._agent_response or self._current_response
-            if resp and resp._has_thinking:
-                if not resp.thinking_section._finalized:
-                    resp.thinking_section.finalize()
-        except RuntimeError:
-            pass  # widget 已被 clear 销毁
-        if self._thinking_timer:
-            self._thinking_timer.stop()
-            self._thinking_timer = None
-        # ★ 停止输入框上方的思考指示条
-        try:
-            self.thinking_bar.stop()
-        except (RuntimeError, AttributeError):
-            pass
-    
-    @QtCore.Slot()
-    def _resume_thinking_main_thread(self):
-        """[主线程] 实际执行恢复思考区块并重启计时器"""
-        if not getattr(self, '_is_running', False):
-            return  # Agent 已停止，忽略延迟到达的信号
-        try:
-            resp = self._agent_response or self._current_response
-            if resp and resp._has_thinking:
-                ts = resp.thinking_section
-                if ts._finalized:
-                    ts.resume()
-        except RuntimeError:
-            pass  # widget 已被 clear 销毁
-        # 重启计时器（如果已停止）
-        if not self._thinking_timer:
-            self._thinking_timer = QtCore.QTimer(self)
-            self._thinking_timer.timeout.connect(lambda: self._updateThinkingTime.emit())
-            self._thinking_timer.start(1000)
-        # ★ 重新启动输入框上方的思考指示条
-        try:
-            self.thinking_bar.start()
-        except (RuntimeError, AttributeError):
-            pass
-
-    def _emit_normal_content(self, text: str):
-        """发送正式内容（带 token 限制 + 自适应缓冲刷新）
-        
-        ★ 自适应策略（借鉴 markstream-vue 的时间预算机制）：
-        - 首个 chunk 立即刷新，消除首字延迟
-        - 后续根据上一次渲染耗时动态调整缓冲大小：
-          渲染快 → 小缓冲、多刷新（流畅感）
-          渲染慢 → 大缓冲、少刷新（避免卡顿）
-        - 换行始终立即刷新（段落边界及时显示）
-        """
-        if not text:
-            return
-        # 首次正式内容到达时，确保思考区块已 finalize（适配 DeepSeek 原生 reasoning_content）
-        # 使用标志位避免从后台线程访问 Qt 控件属性
-        if self._in_think_block is False and getattr(self, '_thinking_needs_finalize', True):
-            self._finalize_thinking()  # 通过信号分派到主线程
-            self._thinking_needs_finalize = False
-
-        # Token 限制仅对正式内容计数
-        if not self._check_output_token_limit(text):
-            if self._output_buffer:
-                self._appendContent.emit(self._output_buffer)
-                self._output_buffer = ""
-            self._appendContent.emit(tr('ai.token_limit'))
-            self._addStatus.emit(tr('ai.token_limit_status'))
-            self.client.request_stop()
-            return
-
-        self._output_buffer += text
-
-        # ★ 自适应缓冲刷新策略
-        should_flush = False
-        current_time = time.time()
-
-        # 初始化自适应状态（首次调用）
-        if not hasattr(self, '_adaptive_buf_size'):
-            self._adaptive_buf_size = 80       # 初始缓冲大小（字符）
-            self._adaptive_interval = 0.15     # 初始兜底间隔（秒）
-            self._last_render_duration = 0.0   # 上次渲染耗时
-            self._flush_count = 0              # flush 计数（性能追踪）
-            self._is_first_content_chunk = True  # 首个 chunk 标志
-
-        # 规则 1: 首个 chunk 立即刷新（消除首字延迟）
-        if self._is_first_content_chunk:
-            should_flush = True
-            self._is_first_content_chunk = False
-        # 规则 2: 缓冲区达到自适应阈值
-        elif len(self._output_buffer) >= self._adaptive_buf_size:
-            should_flush = True
-        # 规则 3: 换行时立即刷新（段落边界及时显示）
-        elif '\n' in text:
-            should_flush = True
-        # 规则 4: 自适应兜底间隔
-        elif current_time - self._last_flush_time > self._adaptive_interval:
-            should_flush = True
-
-        if should_flush and self._output_buffer:
-            flush_start = time.time()
-
-            # 实时过滤伪造的工具调用行
-            buf = self._output_buffer
-            if '[ok]' in buf or '[err]' in buf or '[工具执行结果]' in buf or '[Tool Result]' in buf:
-                lines = buf.split('\n')
-                filtered = []
-                has_fake = False
-                for ln in lines:
-                    s = ln.strip()
-                    if s == '[工具执行结果]' or s == '[Tool Result]' or self._FAKE_TOOL_PATTERNS.match(s):
-                        has_fake = True
-                        continue
-                    filtered.append(ln)
-                buf = '\n'.join(filtered)
-                if has_fake and not getattr(self, '_fake_warned', False):
-                    self._addStatus.emit(tr('ai.fake_tool'))
-                    self._fake_warned = True
-            if buf.strip():
-                self._appendContent.emit(buf)
-            self._output_buffer = ""
-            self._last_flush_time = current_time
-            self._flush_count += 1
-
-            # ★ 自适应调整：根据上次渲染耗时动态调整缓冲参数
-            render_dur = time.time() - flush_start
-            self._last_render_duration = render_dur
-            if render_dur < 0.004:
-                # 渲染很快 → 减小缓冲，更频繁刷新（流畅感）
-                self._adaptive_buf_size = max(40, self._adaptive_buf_size - 20)
-                self._adaptive_interval = max(0.08, self._adaptive_interval - 0.02)
-            elif render_dur > 0.012:
-                # 渲染较慢 → 增大缓冲，减少刷新（避免卡顿）
-                self._adaptive_buf_size = min(500, self._adaptive_buf_size + 40)
-                self._adaptive_interval = min(0.40, self._adaptive_interval + 0.05)
-
-    def _check_output_token_limit(self, text: str) -> bool:
-        """检查正式输出 token 是否超过限制（思考内容不计入）"""
-        if not text:
-            return True
-        new_tokens = self.token_optimizer.estimate_tokens(text)
-        self._current_output_tokens += new_tokens
-        if self._current_output_tokens >= self._max_output_tokens:
-            return False
-        if (self._current_output_tokens >= self._output_token_warning
-                and self._current_output_tokens < self._max_output_tokens):
-            remaining = self._max_output_tokens - self._current_output_tokens
-            if remaining < 400:
-                self._addStatus.emit(
-                    tr('ai.approaching_limit', self._current_output_tokens, self._max_output_tokens))
-        return True
-
-    def _on_thinking_chunk(self, text: str):
-        """处理原生 reasoning_content（DeepSeek R1 等模型）
-        
-        ★ 受 Think 开关控制：关闭时静默丢弃
-        """
-        if text and self._think_enabled:
-            self._addThinking.emit(text)
-    
-    @QtCore.Slot(str)
-    def _on_add_thinking(self, text: str):
-        """在主线程更新思考内容（槽函数）"""
-        if not getattr(self, '_is_running', False):
-            return  # Agent 已停止，忽略延迟到达的信号
-        try:
-            resp = self._agent_response or self._current_response
-            if resp:
-                resp.add_thinking(text)
-                # ★ 首次思考内容 → 启动输入框上方思考指示条
-                if hasattr(self, 'thinking_bar') and not self.thinking_bar.isVisible():
-                    self.thinking_bar.start()
-            self._scroll_agent_to_bottom(force=False)
-        except RuntimeError:
-            pass  # widget 已被 clear 销毁
-
-    def _on_add_status(self, text: str):
-        try:
-            resp = self._agent_response or self._current_response
-            if resp:
-                resp.add_status(text)
-                self._scroll_agent_to_bottom(force=False)
-        except RuntimeError:
-            pass  # widget 已被 clear 销毁
-
-    def _on_update_thinking(self):
-        try:
-            resp = self._agent_response or self._current_response
-            if resp:
-                resp.update_thinking_time()
-                # ★ 同步更新输入框上方思考指示条的时间
-                if hasattr(self, 'thinking_bar') and self.thinking_bar.isVisible():
-                    if resp._has_thinking:
-                        self.thinking_bar.set_elapsed(resp.thinking_section._total_elapsed())
-        except RuntimeError:
-            pass  # 控件可能已销毁
-
     def _cook_displayed_nodes_if_manual(self):
         """★ 在 Manual 保护模式下，对当前工作区的 display 节点做针对性 cook
         
@@ -2410,302 +1467,6 @@ SideFX Labs Node Usage Rules (MUST follow strictly):
     # ------------------------------------------------------------------
     # Plan 模式工具处理
     # ------------------------------------------------------------------
-
-    def _handle_create_plan(self, kwargs: dict) -> dict:
-        """处理 create_plan 工具调用（后台线程）"""
-        try:
-            if self._plan_manager is None:
-                self._plan_manager = get_plan_manager()
-            plan_data = self._plan_manager.create_plan(self._session_id, kwargs)
-            self._plan_phase = 'awaiting_confirmation'
-            # 切换状态：Planning → Generating（Plan 已完成构建）
-            self._showGenerating.emit()
-            # 通过信号在主线程渲染 PlanViewer 卡片
-            self._renderPlanViewer.emit(plan_data)
-            return {
-                "success": True,
-                "result": f"Plan '{plan_data.get('title', '')}' created with {len(plan_data.get('steps', []))} steps. Waiting for user confirmation."
-            }
-        except Exception as e:
-            return {"success": False, "error": f"Failed to create plan: {e}"}
-
-    def _handle_update_plan_step(self, kwargs: dict) -> dict:
-        """处理 update_plan_step 工具调用（后台线程）"""
-        try:
-            if self._plan_manager is None:
-                self._plan_manager = get_plan_manager()
-            step_id = kwargs.get('step_id', '')
-            status = kwargs.get('status', 'done')
-            result_summary = kwargs.get('result_summary', '')
-            plan = self._plan_manager.update_step(
-                self._session_id, step_id, status, result_summary
-            )
-            if not plan:
-                return {"success": False, "error": f"No active plan found for session {self._session_id}"}
-            # 通过信号在主线程更新 PlanViewer 步骤状态
-            self._updatePlanStep.emit(step_id, status, result_summary or '')
-            # 检查是否全部完成
-            all_steps = plan.get('steps', [])
-            done_count = sum(1 for s in all_steps if s.get('status') == 'done')
-            error_count = sum(1 for s in all_steps if s.get('status') == 'error')
-            total = len(all_steps)
-            
-            if plan.get('status') == 'completed':
-                self._plan_phase = 'completed'
-                return {
-                    "success": True,
-                    "result": f"Step {step_id} updated to '{status}'. Plan complete! ({done_count}/{total} done, {error_count} errors)"
-                }
-            
-            # 返回进度信息，让 AI 知道还有多少步骤要做
-            pending_steps = [s for s in all_steps if s.get('status') == 'pending']
-            next_step_info = ""
-            if pending_steps:
-                ns = pending_steps[0]
-                next_step_info = f" Next: {ns['id']} \"{ns.get('title', ns.get('description', ns['id']))}\""
-            
-            return {
-                "success": True,
-                "result": f"Step {step_id} updated to '{status}'. Progress: {done_count}/{total} done.{next_step_info}"
-            }
-        except Exception as e:
-            return {"success": False, "error": f"Failed to update plan step: {e}"}
-
-    def _handle_ask_question(self, kwargs: dict) -> dict:
-        """处理 ask_question 工具调用（后台线程）
-        
-        复用 _request_tool_confirmation 的阻塞模式：
-        1. 设置 pending 属性 → 发射信号 → 主线程渲染 AskQuestionCard
-        2. 后台线程在 queue 上阻塞等待用户回答
-        3. 用户提交后 queue.put(answers) → 后台线程继续
-        """
-        questions = kwargs.get('questions', [])
-        if not questions:
-            return {"success": False, "error": "No questions provided"}
-
-        self._ask_question_result_queue = queue.Queue()
-        self._pending_ask_questions = questions
-        self._askQuestionRequest.emit()
-
-        try:
-            result = self._ask_question_result_queue.get(timeout=300.0)  # 5 分钟超时
-            if result is None:
-                return {"success": True, "result": "User skipped the questions."}
-            # 格式化答案为可读文本
-            answer_lines = []
-            for q_id, selections in result.items():
-                readable = []
-                for sel in selections:
-                    if sel.startswith("__free_text__:"):
-                        readable.append(sel.replace("__free_text__:", ""))
-                    else:
-                        readable.append(sel)
-                answer_lines.append(f"{q_id}: {', '.join(readable)}")
-            return {
-                "success": True,
-                "result": f"User answered:\n" + "\n".join(answer_lines)
-            }
-        except queue.Empty:
-            return {"success": True, "result": "User did not answer within the time limit."}
-
-    @QtCore.Slot()
-    def _on_render_ask_question(self):
-        """主线程：在聊天流中插入 AskQuestionCard"""
-        q = getattr(self, '_ask_question_result_queue', None)
-        questions = getattr(self, '_pending_ask_questions', [])
-
-        if not q:
-            print("[AskQuestion] ⚠ _ask_question_result_queue 不存在")
-            return
-
-        try:
-            card = AskQuestionCard(questions, parent=self.chat_container)
-        except Exception as e:
-            print(f"[AskQuestion] ✖ AskQuestionCard 创建失败: {e}")
-            q.put(None)
-            return
-
-        def _on_answered(answers: dict):
-            q.put(answers)
-
-        def _on_cancelled():
-            q.put(None)
-
-        card.answered.connect(_on_answered)
-        card.cancelled.connect(_on_cancelled)
-
-        # 插入到对话流
-        try:
-            self.chat_layout.insertWidget(self.chat_layout.count() - 1, card)
-        except Exception as e:
-            print(f"[AskQuestion] ⚠ 插入失败: {e}")
-            q.put(None)
-            return
-
-        card.setVisible(True)
-        try:
-            self._scroll_to_bottom(force=True)
-        except Exception:
-            pass
-
-    @QtCore.Slot(dict)
-    def _show_plan_generation_progress(self, accumulated: str):
-        """从 create_plan 的流式参数中提取进度信息并显示 Planning... 状态"""
-        import re as _re
-        # 统计已出现的 step id
-        step_ids = _re.findall(r'"id"\s*:\s*"(step-\d+)"', accumulated)
-        # 尝试提取 title
-        title_match = _re.search(r'"title"\s*:\s*"([^"]{1,30})', accumulated)
-        title_part = title_match.group(1) if title_match else ""
-
-        # 检查是否已进入 architecture 部分
-        has_arch = '"architecture"' in accumulated
-        arch_nodes = _re.findall(r'"id"\s*:\s*"(?!step-)([^"]+)"', accumulated)
-
-        if has_arch and arch_nodes:
-            progress = f"architecture ({len(arch_nodes)} nodes)"
-        elif step_ids:
-            progress = f"step {len(step_ids)}"
-            if title_part:
-                progress = f"「{title_part}」 {progress}"
-        elif title_part:
-            progress = f"「{title_part}」"
-        else:
-            progress = ""
-
-        self._showPlanning.emit(progress)
-
-    @QtCore.Slot()
-    def _on_create_streaming_plan(self):
-        """主线程：创建流式 Plan 预览卡片并插入聊天流"""
-        try:
-            # 如果已有旧的流式卡片则先移除
-            if self._streaming_plan_card is not None:
-                self._streaming_plan_card.setParent(None)
-                self._streaming_plan_card.deleteLater()
-
-            card = StreamingPlanCard(parent=self.chat_container)
-            self._streaming_plan_card = card
-            self.chat_layout.insertWidget(self.chat_layout.count() - 1, card)
-            self._scroll_to_bottom(force=True)
-        except Exception as e:
-            print(f"[Plan] Create streaming card error: {e}")
-
-    @QtCore.Slot(str)
-    def _on_update_streaming_plan(self, accumulated: str):
-        """主线程：将流式 JSON 碎片增量渲染到流式 Plan 卡片
-
-        使用简单的节流策略：缓存最新数据，通过 singleShot 延迟处理，
-        避免每个 token 都触发正则解析和 UI 更新。
-        """
-        self._streaming_plan_acc = accumulated
-        if not getattr(self, '_streaming_plan_timer_active', False):
-            self._streaming_plan_timer_active = True
-            QtCore.QTimer.singleShot(150, self._flush_streaming_plan)
-
-    def _flush_streaming_plan(self):
-        """实际执行流式 Plan 卡片更新"""
-        self._streaming_plan_timer_active = False
-        if self._streaming_plan_card is None:
-            return
-        acc = getattr(self, '_streaming_plan_acc', '')
-        if not acc:
-            return
-        try:
-            old_count = self._streaming_plan_card._rendered_step_count
-            self._streaming_plan_card.update_from_accumulated(acc)
-            new_count = self._streaming_plan_card._rendered_step_count
-            if new_count > old_count:
-                self._scroll_to_bottom()
-        except Exception as e:
-            print(f"[Plan] Update streaming card error: {e}")
-
-    def _on_render_plan_viewer(self, plan_data: dict):
-        """主线程：将流式 Plan 卡片原地升级为完整交互卡片。
-
-        如果流式卡片已存在 → finalize_with_data 原地补充完整数据。
-        如果不存在（边缘情况）→ 创建新卡片。
-        """
-        try:
-            if self._streaming_plan_card is not None:
-                # ★ 原地升级：在流式骨架上补充 DAG + 按钮
-                card = self._streaming_plan_card
-                card.finalize_with_data(plan_data)
-                card.planConfirmed.connect(self._on_plan_confirmed)
-                card.planRejected.connect(self._on_plan_rejected)
-                self._active_plan_viewer = card
-                self._streaming_plan_card = None  # 不再追踪为流式卡片
-            else:
-                # 边缘情况：没有流式卡片时直接创建 PlanViewer
-                viewer = PlanViewer(plan_data, parent=self.chat_container)
-                viewer.planConfirmed.connect(self._on_plan_confirmed)
-                viewer.planRejected.connect(self._on_plan_rejected)
-                self._active_plan_viewer = viewer
-                self.chat_layout.insertWidget(self.chat_layout.count() - 1, viewer)
-            self._scroll_to_bottom(force=True)
-        except Exception as e:
-            print(f"[Plan] Render PlanViewer error: {e}")
-
-    @QtCore.Slot(str, str, str)
-    def _on_update_plan_step(self, step_id: str, status: str, result_summary: str):
-        """主线程：更新 PlanViewer 卡片中的步骤状态"""
-        if self._active_plan_viewer:
-            try:
-                self._active_plan_viewer.update_step_status(step_id, status, result_summary)
-            except Exception as e:
-                print(f"[Plan] Update step UI error: {e}")
-
-    def _on_plan_confirmed(self, plan_data: dict):
-        """用户点击 Confirm 按钮 → 启动执行阶段"""
-        self._plan_phase = 'executing'
-        # 禁用 PlanViewer 按钮（防止重复点击）
-        if self._active_plan_viewer:
-            self._active_plan_viewer.set_confirmed()
-        
-        # 构造执行提示消息
-        exec_msg = tr('ai.plan_confirmed_msg', plan_data.get('title', 'Plan'))
-        _pmsg = {'role': 'user', 'content': exec_msg}
-        self._conversation_history.append(_pmsg)
-        if self._send_context is not None:
-            self._send_context.append(_pmsg)
-        
-        # 创建新的 AI 回复块
-        self._set_running(True)
-        self._add_ai_response()
-        self._agent_response = self._current_response
-        self._start_active_aurora()
-        
-        # 构造 agent_params（复用上次的 provider/model 设置）
-        agent_params = getattr(self, '_last_agent_params', {}).copy()
-        agent_params['use_agent'] = True          # 执行阶段用完整工具
-        agent_params['plan_mode'] = True
-        agent_params['plan_executing'] = True     # 标记为 Plan 执行阶段
-        agent_params['plan_data'] = plan_data
-        
-        # 后台线程执行
-        thread = threading.Thread(
-            target=self._run_agent, args=(agent_params,), daemon=True
-        )
-        thread.start()
-
-    def _on_plan_rejected(self):
-        """用户点击 Reject 按钮 → 丢弃 Plan"""
-        self._plan_phase = 'idle'
-        try:
-            if self._plan_manager is None:
-                self._plan_manager = get_plan_manager()
-            self._plan_manager.delete_plan(self._session_id)
-        except Exception:
-            pass
-        if self._active_plan_viewer:
-            self._active_plan_viewer.set_rejected()
-        self._active_plan_viewer = None
-
-    # 已自带 checkpoint 追踪的工具（在 _on_add_node_operation 中有专用分支）
-    _SELF_TRACKING_TOOLS = frozenset({
-        'create_node', 'create_nodes_batch', 'create_wrangle_node',
-        'delete_node', 'set_node_parameter',
-    })
 
     @staticmethod
     def _snapshot_network_children() -> dict:
@@ -5551,235 +4312,6 @@ SideFX Labs Node Usage Rules (MUST follow strictly):
     # 图片输入支持
     # ============================================================
     
-    def _current_model_supports_vision(self) -> bool:
-        """检查当前选中的模型是否支持图片输入"""
-        model = self.model_combo.currentText()
-        features = self._model_features.get(model, {})
-        return features.get('supports_vision', False)
-    
-    def _on_attach_image(self):
-        """打开文件对话框选择图片"""
-        if not self._current_model_supports_vision():
-            model = self.model_combo.currentText()
-            QtWidgets.QMessageBox.information(
-                self, "不支持图片",
-                f"当前模型 {model} 不支持图片输入。\n请切换到支持视觉的模型（如 Claude、GPT-5.2 等）。"
-            )
-            return
-        
-        file_paths, _ = QtWidgets.QFileDialog.getOpenFileNames(
-            self, "选择图片", "",
-            "Images (*.png *.jpg *.jpeg *.gif *.webp *.bmp);;All Files (*)"
-        )
-        for fp in file_paths:
-            self._add_image_from_path(fp)
-    
-    def _add_image_from_path(self, file_path: str):
-        """从文件路径加载图片并添加到待发送列表（自动缩放过大图片）"""
-        import base64
-        try:
-            # ★ 通过 QImage 加载，统一走缩放逻辑
-            qimg = QtGui.QImage(file_path)
-            if qimg.isNull():
-                print(f"[AI Tab] 无法加载图片: {file_path}")
-                return
-            qimg = self._resize_image_if_needed(qimg, self._MAX_IMAGE_DIMENSION)
-            
-            ext = os.path.splitext(file_path)[1].lower()
-            # 优先保持原始格式；BMP/GIF 等不适合直接发 API，统一转 PNG
-            if ext in ('.jpg', '.jpeg'):
-                fmt, media_type = 'JPEG', 'image/jpeg'
-            elif ext == '.webp':
-                fmt, media_type = 'WEBP', 'image/webp'
-            else:
-                fmt, media_type = 'PNG', 'image/png'
-            
-            buf = QtCore.QBuffer()
-            buf.open(QtCore.QIODevice.WriteOnly)
-            quality = 90 if fmt == 'JPEG' else -1
-            qimg.save(buf, fmt, quality)
-            raw_bytes = buf.data().data()
-            buf.close()
-            
-            # ★ 过大时降级为 JPEG 压缩
-            if len(raw_bytes) > self._MAX_IMAGE_BYTES and fmt != 'JPEG':
-                buf2 = QtCore.QBuffer()
-                buf2.open(QtCore.QIODevice.WriteOnly)
-                qimg.save(buf2, 'JPEG', 85)
-                raw_bytes = buf2.data().data()
-                buf2.close()
-                media_type = 'image/jpeg'
-                print(f"[AI Tab] 图片过大，已转为 JPEG ({len(raw_bytes)//1024}KB)")
-            
-            b64 = base64.b64encode(raw_bytes).decode('utf-8')
-            self._add_pending_image(b64, media_type)
-        except Exception as e:
-            print(f"[AI Tab] 加载图片失败: {e}")
-    
-    # ★ 图片尺寸限制：超过此分辨率的图片自动缩放（防止 base64 过大导致 API 400 错误）
-    _MAX_IMAGE_DIMENSION = 2048  # 最长边不超过 2048px
-    _MAX_IMAGE_BYTES = 5 * 1024 * 1024  # base64 前的原始字节数上限 ~5MB（编码后约 6.7MB）
-
-    @staticmethod
-    def _resize_image_if_needed(image: 'QtGui.QImage', max_dim: int = 2048) -> 'QtGui.QImage':
-        """如果图片超过 max_dim，等比缩放。返回缩放后的 QImage。"""
-        w, h = image.width(), image.height()
-        if w <= max_dim and h <= max_dim:
-            return image
-        if w > h:
-            new_w = max_dim
-            new_h = int(h * max_dim / w)
-        else:
-            new_h = max_dim
-            new_w = int(w * max_dim / h)
-        print(f"[AI Tab] 图片过大 ({w}x{h})，自动缩放至 {new_w}x{new_h}")
-        return image.scaled(new_w, new_h, QtCore.Qt.KeepAspectRatio, QtCore.Qt.SmoothTransformation)
-
-    def _on_image_dropped(self, image: 'QtGui.QImage'):
-        """ChatInput 拖拽或粘贴图片的回调"""
-        if not self._current_model_supports_vision():
-            return
-        import base64
-        # ★ 自动缩放过大图片
-        image = self._resize_image_if_needed(image, self._MAX_IMAGE_DIMENSION)
-        buf = QtCore.QBuffer()
-        buf.open(QtCore.QIODevice.WriteOnly)
-        image.save(buf, "PNG")
-        raw_bytes = buf.data().data()
-        buf.close()
-        # ★ 如果 PNG 仍然过大，改用 JPEG 压缩
-        if len(raw_bytes) > self._MAX_IMAGE_BYTES:
-            buf2 = QtCore.QBuffer()
-            buf2.open(QtCore.QIODevice.WriteOnly)
-            image.save(buf2, "JPEG", 85)
-            raw_bytes = buf2.data().data()
-            buf2.close()
-            media_type = 'image/jpeg'
-            print(f"[AI Tab] PNG 过大，已转为 JPEG (quality=85, {len(raw_bytes)//1024}KB)")
-        else:
-            media_type = 'image/png'
-        b64 = base64.b64encode(raw_bytes).decode('utf-8')
-        self._add_pending_image(b64, media_type)
-    
-    def _add_pending_image(self, b64_data: str, media_type: str):
-        """添加图片到待发送列表并在预览区显示缩略图（点击可放大）"""
-        # 创建缩略图和完整 pixmap
-        img_bytes = __import__('base64').b64decode(b64_data)
-        full_pixmap = QtGui.QPixmap()
-        full_pixmap.loadFromData(img_bytes)
-        thumb = full_pixmap.scaled(60, 60, QtCore.Qt.KeepAspectRatio, QtCore.Qt.SmoothTransformation)
-        
-        # 存储
-        idx = len(self._pending_images)
-        self._pending_images.append((b64_data, media_type, thumb))
-        
-        # 创建预览 widget
-        img_widget = QtWidgets.QWidget()
-        img_layout = QtWidgets.QVBoxLayout(img_widget)
-        img_layout.setContentsMargins(2, 2, 2, 2)
-        img_layout.setSpacing(1)
-        
-        lbl = ClickableImageLabel(thumb, full_pixmap)
-        lbl.setObjectName("imgThumb")
-        img_layout.addWidget(lbl)
-        
-        # 删除按钮
-        rm_btn = QtWidgets.QPushButton("x")
-        rm_btn.setFixedSize(16, 16)
-        rm_btn.setObjectName("imgRemoveBtn")
-        rm_btn.clicked.connect(lambda checked=False, i=idx: self._remove_pending_image(i))
-        img_layout.addWidget(rm_btn, alignment=QtCore.Qt.AlignCenter)
-        
-        # 插入到 stretch 之前
-        count = self.image_preview_layout.count()
-        self.image_preview_layout.insertWidget(count - 1, img_widget)
-        self.image_preview_container.setVisible(True)
-    
-    def _remove_pending_image(self, index: int):
-        """移除待发送图片"""
-        if 0 <= index < len(self._pending_images):
-            self._pending_images[index] = None  # 标记为已删除
-            self._rebuild_image_preview()  # 过滤 None 后重建整个预览区
-    
-    def _rebuild_image_preview(self):
-        """重新构建图片预览区"""
-        # 清除所有 widget（保留 stretch）
-        while self.image_preview_layout.count() > 1:
-            item = self.image_preview_layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-        
-        # 重新过滤并添加
-        new_images = [(b64, mt, th) for entry in self._pending_images 
-                      if entry is not None for b64, mt, th in [entry]]
-        self._pending_images = list(new_images)
-        
-        if not self._pending_images:
-            self.image_preview_container.setVisible(False)
-            return
-        
-        for i, (b64, mt, thumb) in enumerate(self._pending_images):
-            img_widget = QtWidgets.QWidget()
-            img_layout = QtWidgets.QVBoxLayout(img_widget)
-            img_layout.setContentsMargins(2, 2, 2, 2)
-            img_layout.setSpacing(1)
-            
-            # 从 base64 还原完整 pixmap 用于放大预览
-            full_pixmap = QtGui.QPixmap()
-            full_pixmap.loadFromData(__import__('base64').b64decode(b64))
-            lbl = ClickableImageLabel(thumb, full_pixmap)
-            lbl.setObjectName("imgThumb")
-            img_layout.addWidget(lbl)
-            
-            rm_btn = QtWidgets.QPushButton("x")
-            rm_btn.setFixedSize(16, 16)
-            rm_btn.setObjectName("imgRemoveBtn")
-            rm_btn.clicked.connect(lambda checked=False, idx=i: self._remove_pending_image(idx))
-            img_layout.addWidget(rm_btn, alignment=QtCore.Qt.AlignCenter)
-            
-            count = self.image_preview_layout.count()
-            self.image_preview_layout.insertWidget(count - 1, img_widget)
-    
-    def _clear_pending_images(self):
-        """清空所有待发送图片"""
-        self._pending_images.clear()
-        while self.image_preview_layout.count() > 1:
-            item = self.image_preview_layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-        self.image_preview_container.setVisible(False)
-    
-    def _build_multimodal_content(self, text: str, images: list) -> list:
-        """构建包含文字和图片的多模态消息内容（OpenAI Vision API 格式）
-        
-        Args:
-            text: 用户文字消息
-            images: List of (base64_data, media_type, thumbnail) tuples
-            
-        Returns:
-            list: content 数组，包含 text 和 image_url 项
-        """
-        # ★ API 支持的 media type 白名单（BMP 等需要先转换）
-        _SUPPORTED_MEDIA = {'image/png', 'image/jpeg', 'image/gif', 'image/webp'}
-        
-        content_parts = []
-        # ★ 始终添加 text 部分（即使为空也提供占位符，某些 API 要求至少一个 text block）
-        content_parts.append({"type": "text", "text": text or " "})
-        # 添加图片
-        for b64_data, media_type, _thumb in images:
-            if not b64_data:
-                continue  # 跳过空数据
-            # ★ 不支持的 media type 降级为 image/png
-            if media_type not in _SUPPORTED_MEDIA:
-                media_type = 'image/png'
-            content_parts.append({
-                "type": "image_url",
-                "image_url": {
-                    "url": f"data:{media_type};base64,{b64_data}"
-                }
-            })
-        return content_parts
-    
     def _on_read_selection(self):
         ok, text = self.mcp.describe_selection()
         if ok:
@@ -5942,1986 +4474,4 @@ SideFX Labs Node Usage Rules (MUST follow strictly):
             QtWidgets.QMessageBox.critical(self, "导出错误", f"导出训练数据时发生错误：{str(e)}")
 
     # ===== 缓存管理 =====
-    
-    def _on_cache_menu(self):
-        """显示缓存菜单"""
-        menu = QtWidgets.QMenu(self)
-        
-        # 保存存档（独立文件）
-        archive_action = menu.addAction("存档当前对话")
-        archive_action.triggered.connect(self._archive_cache)
-        
-        # 加载对话
-        load_action = menu.addAction("加载对话...")
-        load_action.triggered.connect(self._load_cache_dialog)
-        
-        menu.addSeparator()
-        
-        # 压缩为摘要（减少 token）
-        compress_action = menu.addAction("压缩旧对话为摘要")
-        compress_action.triggered.connect(self._compress_to_summary)
-        
-        # 列出所有缓存
-        list_action = menu.addAction("查看所有缓存")
-        list_action.triggered.connect(self._list_caches)
-        
-        menu.addSeparator()
-        
-        # 自动保存开关
-        auto_save_action = menu.addAction("[on] 自动保存" if self._auto_save_cache else "自动保存")
-        auto_save_action.setCheckable(True)
-        auto_save_action.setChecked(self._auto_save_cache)
-        auto_save_action.triggered.connect(lambda: setattr(self, '_auto_save_cache', not self._auto_save_cache))
-        
-        # 显示菜单
-        menu.exec_(self.btn_cache.mapToGlobal(QtCore.QPoint(0, self.btn_cache.height())))
-    
-    @staticmethod
-    def _strip_images_for_cache(history: list) -> list:
-        """剥离 conversation_history 中的 base64 图片数据，
-        用占位文本替代，大幅减小缓存文件体积。
-        返回一份深拷贝，不修改原始 history。
-        """
-        import copy
-        stripped = []
-        for msg in history:
-            content = msg.get('content')
-            if isinstance(content, list):
-                # 多模态消息：content 是 [{type:text,...}, {type:image_url,...}, ...]
-                new_parts = []
-                for part in content:
-                    if part.get('type') == 'image_url':
-                        url = part.get('image_url', {}).get('url', '')
-                        if url.startswith('data:'):
-                            # 替换 base64 为占位符，保留 media type 信息
-                            media_type = url.split(';')[0].replace('data:', '')
-                            new_parts.append({
-                                'type': 'text',
-                                'text': f'[Image: {media_type}]',
-                            })
-                        else:
-                            new_parts.append(copy.copy(part))
-                    else:
-                        new_parts.append(copy.copy(part))
-                new_msg = msg.copy()
-                new_msg['content'] = new_parts
-                stripped.append(new_msg)
-            else:
-                stripped.append(msg)  # 非多模态消息直接引用（str/None 不可变）
-        return stripped
-    
-    def _build_cache_data(self) -> dict:
-        """构建缓存数据字典"""
-        todo_data = []
-        if hasattr(self, 'todo_list') and self.todo_list:
-            todo_data = self.todo_list.get_todos_data()
-        return {
-            'version': '1.0',
-            'session_id': self._session_id,
-            'created_at': datetime.now().isoformat(),
-            'message_count': len(self._conversation_history),
-            'estimated_tokens': self._calculate_context_tokens(),
-            'conversation_history': self._conversation_history,
-            'context_summary': self._context_summary,
-            'todo_summary': self.todo_list.get_todos_summary() if hasattr(self, 'todo_list') else "",
-            'todo_data': todo_data,
-            'token_stats': self._token_stats.copy(),
-        }
-
-    def _on_destroyed(self):
-        """Widget 被销毁时标记，防止旧实例的 atexit/aboutToQuit 回调覆盖新数据"""
-        self._destroyed = True
-        try:
-            app = QtWidgets.QApplication.instance()
-            if app:
-                try:
-                    app.aboutToQuit.disconnect(self._save_all_sessions)
-                except (TypeError, RuntimeError):
-                    pass
-        except Exception:
-            pass
-
-    def _periodic_save_all(self):
-        """定期保存所有会话（QTimer 触发 + aboutToQuit 触发）"""
-        try:
-            if not self._sessions:
-                return
-            # 只有存在对话时才保存
-            has_any = False
-            for sid, sdata in self._sessions.items():
-                if sdata.get('conversation_history'):
-                    has_any = True
-                    break
-            if not has_any:
-                return
-            self._save_all_sessions()
-        except Exception as e:
-            logger.warning("[Cache] 定期保存失败: %s", e)
-    
-    def _atexit_save(self):
-        """Python 退出时的最后保存机会（atexit 回调）
-        
-        ★ 此时 Qt widget 可能已被销毁，因此：
-        - 使用 _tabs_backup（纯 Python 列表）代替遍历 QTabBar
-        - 使用 try/except 包裹 todo_list 访问
-        - 如果 aboutToQuit 已成功保存过，则跳过（避免用不完整数据覆盖）
-        - 如果 widget 已被销毁（旧实例），跳过以免覆盖新实例的数据
-        """
-        try:
-            if getattr(self, '_destroyed', False):
-                return
-            if getattr(self, '_sessions_saved', False):
-                return
-            if not hasattr(self, '_sessions') or not self._sessions:
-                return
-            logger.debug("[Cache] atexit: 开始保存 (sessions=%d, backup=%d)", len(self._sessions), len(getattr(self, '_tabs_backup', [])))
-            try:
-                self._save_current_session_state()
-            except (RuntimeError, AttributeError):
-                pass
-            
-            tabs_info = getattr(self, '_tabs_backup', [])
-            if not tabs_info:
-                tabs_info = [(sid, f"Chat") for sid in self._sessions]
-            
-            manifest_tabs = []
-            for sid, tab_label in tabs_info:
-                if not sid or sid not in self._sessions:
-                    continue
-                sdata = self._sessions[sid]
-                history = sdata.get('conversation_history', [])
-                if not history:
-                    manifest_tabs.append({
-                        'session_id': sid,
-                        'tab_label': tab_label,
-                        'file': '',
-                        'empty': True,
-                    })
-                    continue
-                todo_data = []
-                try:
-                    todo_list_obj = sdata.get('todo_list')
-                    todo_data = todo_list_obj.get_todos_data() if todo_list_obj else []
-                except (RuntimeError, AttributeError, Exception):
-                    pass
-                cache_data = {
-                    'version': '1.0',
-                    'session_id': sid,
-                    'message_count': len(history),
-                    'conversation_history': self._strip_images_for_cache(history),
-                    'context_summary': sdata.get('context_summary', ''),
-                    'todo_data': todo_data,
-                    'token_stats': sdata.get('token_stats', {}),
-                }
-                session_file = self._cache_dir / f"session_{sid}.json"
-                with open(session_file, 'w', encoding='utf-8') as f:
-                    json.dump(cache_data, f, ensure_ascii=False)
-                manifest_tabs.append({
-                    'session_id': sid,
-                    'tab_label': tab_label,
-                    'file': f"session_{sid}.json",
-                })
-            if not manifest_tabs:
-                return
-            # 防止用更少的 tab 数据覆盖已有的完整 manifest
-            manifest_file = self._cache_dir / "sessions_manifest.json"
-            try:
-                if manifest_file.exists():
-                    import json as _json
-                    with open(manifest_file, 'r', encoding='utf-8') as f:
-                        existing = _json.load(f)
-                    existing_count = len(existing.get('tabs', []))
-                    if existing_count > len(manifest_tabs):
-                        return
-            except Exception:
-                pass
-            manifest = {
-                'version': '1.0',
-                'active_session_id': self._session_id,
-                'tabs': manifest_tabs,
-            }
-            with open(manifest_file, 'w', encoding='utf-8') as f:
-                json.dump(manifest, f, ensure_ascii=False)
-        except Exception:
-            pass
-
-    def _save_cache(self) -> bool:
-        """自动保存：覆写同 session 文件 + manifest"""
-        if not self._conversation_history:
-            return False
-        try:
-            # 同步当前会话状态到 _sessions
-            self._save_current_session_state()
-            # ★ 同步 tab 备份
-            self._sync_tabs_backup()
-            
-            cache_data = self._build_cache_data()
-            # ★ 剥离 base64 图片以减小缓存文件大小
-            cache_data['conversation_history'] = self._strip_images_for_cache(
-                cache_data.get('conversation_history', [])
-            )
-
-            # 1. 覆写固定的 session 文件（一个 session 只有一个文件）
-            session_file = self._cache_dir / f"session_{self._session_id}.json"
-            with open(session_file, 'w', encoding='utf-8') as f:
-                json.dump(cache_data, f, ensure_ascii=False, indent=2)
-
-            # 2. 同步更新 sessions_manifest.json（确保所有 tab 信息都是最新的）
-            # ★ 不再写 cache_latest.json — 恢复由 sessions_manifest + session_*.json 管理
-            self._update_manifest()
-
-            if self._workspace_dir:
-                self._update_workspace_cache_info()
-            return True
-        except Exception as e:
-            logger.warning("[Cache] 自动保存失败: %s", e)
-            return False
-    
-    def _update_manifest(self):
-        """更新 sessions_manifest.json 以反映当前所有标签的状态"""
-        try:
-            manifest_tabs = []
-            for i in range(self.session_tabs.count()):
-                sid = self.session_tabs.tabData(i)
-                if not sid:
-                    continue
-                tab_label = self.session_tabs.tabText(i)
-                session_file = self._cache_dir / f"session_{sid}.json"
-                if session_file.exists():
-                    manifest_tabs.append({
-                        'session_id': sid,
-                        'tab_label': tab_label,
-                        'file': f"session_{sid}.json",
-                    })
-                else:
-                    sdata = self._sessions.get(sid, {})
-                    history = sdata.get('conversation_history', [])
-                    if history:
-                        manifest_tabs.append({
-                            'session_id': sid,
-                            'tab_label': tab_label,
-                            'file': f"session_{sid}.json",
-                        })
-                    else:
-                        manifest_tabs.append({
-                            'session_id': sid,
-                            'tab_label': tab_label,
-                            'file': '',
-                            'empty': True,
-                        })
-            if manifest_tabs:
-                manifest = {
-                    'version': '1.0',
-                    'active_session_id': self._session_id,
-                    'tabs': manifest_tabs,
-                }
-                manifest_file = self._cache_dir / "sessions_manifest.json"
-                with open(manifest_file, 'w', encoding='utf-8') as f:
-                    json.dump(manifest, f, ensure_ascii=False, indent=2)
-        except Exception as e:
-            logger.warning("[Cache] 更新 manifest 失败: %s", e)
-
-    def _save_all_sessions(self) -> bool:
-        """保存所有打开的会话到磁盘（关闭软件时调用）"""
-        if getattr(self, '_destroyed', False):
-            return False
-        try:
-            # 先保存当前活跃会话的状态到 _sessions 字典
-            try:
-                self._save_current_session_state()
-            except (RuntimeError, AttributeError):
-                pass
-            # ★ 同步 tab 备份（确保 atexit 时也能用）
-            try:
-                self._sync_tabs_backup()
-            except (RuntimeError, AttributeError):
-                pass
-
-            manifest_tabs = []
-            active_session_id = self._session_id
-
-            # 从 QTabBar 获取 tab 列表；如果 Qt widget 已销毁则回退到纯 Python 备份
-            tabs_list = []
-            try:
-                for i in range(self.session_tabs.count()):
-                    sid = self.session_tabs.tabData(i)
-                    tab_label = self.session_tabs.tabText(i)
-                    if sid:
-                        tabs_list.append((sid, tab_label))
-            except (RuntimeError, AttributeError):
-                tabs_list = getattr(self, '_tabs_backup', [])
-
-            for sid, tab_label in tabs_list:
-                if not sid or sid not in self._sessions:
-                    continue
-
-                sdata = self._sessions[sid]
-                history = sdata.get('conversation_history', [])
-                if not history:
-                    # 空会话：清理磁盘残留，但仍记录到 manifest 以保留标签布局
-                    try:
-                        old_file = self._cache_dir / f"session_{sid}.json"
-                        if old_file.exists():
-                            old_file.unlink()
-                    except Exception:
-                        pass
-                    manifest_tabs.append({
-                        'session_id': sid,
-                        'tab_label': tab_label,
-                        'file': '',
-                        'empty': True,
-                    })
-                    continue
-
-                # 收集 todo 数据（防御 widget 已销毁的情况）
-                todo_data = []
-                try:
-                    todo_list_obj = sdata.get('todo_list')
-                    todo_data = todo_list_obj.get_todos_data() if todo_list_obj else []
-                except (RuntimeError, AttributeError):
-                    pass
-
-                # 写 session 文件（★ 剥离 base64 图片以减小文件大小）
-                cache_data = {
-                    'version': '1.0',
-                    'session_id': sid,
-                    'created_at': datetime.now().isoformat(),
-                    'message_count': len(history),
-                    'conversation_history': self._strip_images_for_cache(history),
-                    'context_summary': sdata.get('context_summary', ''),
-                    'todo_data': todo_data,
-                    'token_stats': sdata.get('token_stats', {}),
-                }
-                session_file = self._cache_dir / f"session_{sid}.json"
-                with open(session_file, 'w', encoding='utf-8') as f:
-                    json.dump(cache_data, f, ensure_ascii=False, indent=2)
-
-                manifest_tabs.append({
-                    'session_id': sid,
-                    'tab_label': tab_label,
-                    'file': f"session_{sid}.json",
-                })
-
-            if not manifest_tabs:
-                return False
-
-            manifest = {
-                'version': '1.0',
-                'active_session_id': active_session_id,
-                'tabs': manifest_tabs,
-            }
-            manifest_file = self._cache_dir / "sessions_manifest.json"
-            with open(manifest_file, 'w', encoding='utf-8') as f:
-                json.dump(manifest, f, ensure_ascii=False, indent=2)
-
-            self._sessions_saved = True
-            logger.debug("[Cache] 已保存 %d 个会话标签 (tabs_list=%d, sessions=%d)", len(manifest_tabs), len(tabs_list), len(self._sessions))
-            return True
-        except Exception:
-            logger.exception("[Cache] 保存所有会话失败")
-            return False
-
-    def _restore_all_sessions(self) -> bool:
-        """从 sessions_manifest.json 恢复所有会话标签（启动时调用，幂等）"""
-        # ★ 幂等保护：防止 __init__ 和 main_window 延迟回调重复恢复
-        if getattr(self, '_sessions_restored', False):
-            return True
-        try:
-            manifest_file = self._cache_dir / "sessions_manifest.json"
-            if not manifest_file.exists():
-                return False
-
-            with open(manifest_file, 'r', encoding='utf-8') as f:
-                manifest = json.load(f)
-
-            tabs_info = manifest.get('tabs', [])
-            if not tabs_info:
-                return False
-
-            active_sid = manifest.get('active_session_id', '')
-            active_tab_index = 0
-            first_tab = True
-
-            for tab_info in tabs_info:
-                sid = tab_info.get('session_id', '')
-                tab_label = tab_info.get('tab_label', 'Chat')
-                is_empty = tab_info.get('empty', False)
-
-                history = []
-                context_summary = ''
-                todo_data = []
-                saved_token_stats = {
-                    'input_tokens': 0, 'output_tokens': 0,
-                    'reasoning_tokens': 0,
-                    'cache_read': 0, 'cache_write': 0,
-                    'total_tokens': 0, 'requests': 0,
-                    'estimated_cost': 0.0,
-                }
-
-                if not is_empty:
-                    file_name = tab_info.get('file', '')
-                    if not file_name:
-                        continue
-                    session_file = self._cache_dir / file_name
-                    if not session_file.exists():
-                        continue
-                    with open(session_file, 'r', encoding='utf-8') as f:
-                        cache_data = json.load(f)
-                    history = cache_data.get('conversation_history', [])
-                    if not history:
-                        is_empty = True
-                    else:
-                        context_summary = cache_data.get('context_summary', '')
-                        todo_data = cache_data.get('todo_data', [])
-                        saved_token_stats = cache_data.get('token_stats', saved_token_stats)
-
-                if first_tab:
-                    # 第一个 tab：加载到已有的初始会话中
-                    first_tab = False
-                    old_id = self._session_id
-
-                    self._session_id = sid
-                    self._conversation_history = history
-                    self._context_summary = context_summary
-                    self._token_stats = saved_token_stats
-
-                    if old_id in self._sessions:
-                        sdata = self._sessions.pop(old_id)
-                        sdata['conversation_history'] = history
-                        sdata['context_summary'] = context_summary
-                        sdata['token_stats'] = saved_token_stats
-                        self._sessions[sid] = sdata
-                    elif sid not in self._sessions:
-                        self._sessions[sid] = {
-                            'scroll_area': self.scroll_area,
-                            'chat_container': self.chat_container,
-                            'chat_layout': self.chat_layout,
-                            'todo_list': self.todo_list,
-                            'conversation_history': history,
-                            'context_summary': context_summary,
-                            'current_response': None,
-                            'token_stats': saved_token_stats,
-                        }
-
-                    if todo_data and hasattr(self, 'todo_list') and self.todo_list:
-                        self.todo_list.restore_todos(todo_data)
-                        self._ensure_todo_in_chat(self.todo_list, self.chat_layout)
-
-                    for i in range(self.session_tabs.count()):
-                        if self.session_tabs.tabData(i) == old_id:
-                            self.session_tabs.setTabData(i, sid)
-                            self.session_tabs.setTabText(i, tab_label)
-                            if sid == active_sid:
-                                active_tab_index = i
-                            break
-
-                    if not is_empty:
-                        self._render_conversation_history()
-                else:
-                    # 后续 tab：创建新标签
-                    self._save_current_session_state()
-                    self._session_counter += 1
-
-                    scroll_area, chat_container, chat_layout = self._create_session_widgets()
-                    self.session_stack.addWidget(scroll_area)
-
-                    tab_index = self.session_tabs.addTab(tab_label)
-                    self.session_tabs.setTabData(tab_index, sid)
-
-                    todo = self._create_todo_list(chat_container)
-                    if todo_data:
-                        todo.restore_todos(todo_data)
-                        self._ensure_todo_in_chat(todo, chat_layout)
-
-                    self._sessions[sid] = {
-                        'scroll_area': scroll_area,
-                        'chat_container': chat_container,
-                        'chat_layout': chat_layout,
-                        'todo_list': todo,
-                        'conversation_history': history,
-                        'context_summary': context_summary,
-                        'current_response': None,
-                        'token_stats': saved_token_stats,
-                    }
-
-                    if not is_empty:
-                        # 临时切换到该标签以渲染历史
-                        old_scroll = self.scroll_area
-                        old_chat_container = self.chat_container
-                        old_chat_layout = self.chat_layout
-                        old_todo = self.todo_list
-                        old_history = self._conversation_history
-                        old_summary = self._context_summary
-                        old_stats = self._token_stats
-                        old_sid = self._session_id
-
-                        self._session_id = sid
-                        self._conversation_history = history
-                        self._context_summary = context_summary
-                        self._token_stats = saved_token_stats
-                        self.scroll_area = scroll_area
-                        self.chat_container = chat_container
-                        self.chat_layout = chat_layout
-                        self.todo_list = todo
-
-                        self._render_conversation_history()
-
-                        self._session_id = old_sid
-                        self._conversation_history = old_history
-                        self._context_summary = old_summary
-                        self._token_stats = old_stats
-                        self.scroll_area = old_scroll
-                        self.chat_container = old_chat_container
-                        self.chat_layout = old_chat_layout
-                        self.todo_list = old_todo
-
-                    if sid == active_sid:
-                        active_tab_index = tab_index
-
-            # 切换到之前活跃的标签
-            if self.session_tabs.count() > 0:
-                self.session_tabs.blockSignals(True)
-                self.session_tabs.setCurrentIndex(active_tab_index)
-                self.session_tabs.blockSignals(False)
-
-                target_sid = self.session_tabs.tabData(active_tab_index)
-                if target_sid and target_sid in self._sessions:
-                    self._load_session_state(target_sid)
-                    self.session_stack.setCurrentWidget(
-                        self._sessions[target_sid]['scroll_area']
-                    )
-
-            # ★ 恢复完成后同步 tab 备份并更新 UI 显示
-            self._sync_tabs_backup()
-            self._update_token_stats_display()
-            self._update_context_stats()
-            self._sessions_restored = True  # 标记已恢复，防止重复
-            logger.debug("[Cache] 已恢复 %d 个会话标签", self.session_tabs.count())
-            return True
-
-        except Exception:
-            logger.exception("[Cache] 恢复多会话失败")
-            return False
-
-    def _archive_cache(self) -> bool:
-        """手动存档：创建带时间戳的独立文件（不会被覆写）"""
-        if not self._conversation_history:
-            QtWidgets.QMessageBox.information(self, "提示", "没有对话历史可存档")
-            return False
-        try:
-            cache_data = self._build_cache_data()
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            filename = f"archive_{self._session_id}_{timestamp}.json"
-            archive_file = self._cache_dir / filename
-            with open(archive_file, 'w', encoding='utf-8') as f:
-                json.dump(cache_data, f, ensure_ascii=False, indent=2)
-            est = cache_data['estimated_tokens']
-            self._addStatus.emit(f"已存档: {filename} (~{est} tokens)")
-            return True
-        except Exception as e:
-            QtWidgets.QMessageBox.warning(self, "错误", f"存档失败: {str(e)}")
-            return False
-    
-    def _update_workspace_cache_info(self):
-        """更新工作区中的缓存信息（供主窗口保存工作区时使用）"""
-        # 这个方法会被主窗口调用，用于更新工作区配置
-        # 实际保存由主窗口的 _save_workspace 完成
-        pass
-    
-    def _load_cache(self, cache_file: Path, silent: bool = False) -> bool:
-        """从缓存文件加载对话历史（在新标签页中打开）
-        
-        Args:
-            cache_file: 缓存文件路径
-            silent: 是否静默加载（不显示确认对话框，用于工作区自动恢复）
-        """
-        try:
-            with open(cache_file, 'r', encoding='utf-8') as f:
-                cache_data = json.load(f)
-            
-            # 验证数据格式
-            if 'conversation_history' not in cache_data:
-                if not silent:
-                    QtWidgets.QMessageBox.warning(self, "错误", "缓存文件格式无效")
-                return False
-            
-            # 确认加载（静默模式下跳过）
-            if not silent:
-                msg_count = len(cache_data.get('conversation_history', []))
-                reply = QtWidgets.QMessageBox.question(
-                    self, "确认加载",
-                    f"将在新标签页加载 {msg_count} 条对话记录。\n是否继续？",
-                    QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No
-                )
-                
-                if reply != QtWidgets.QMessageBox.Yes:
-                    return False
-            
-            history = cache_data.get('conversation_history', [])
-            context_summary = cache_data.get('context_summary', '')
-            todo_data = cache_data.get('todo_data', [])
-            cached_session_id = cache_data.get('session_id', str(uuid.uuid4())[:8])
-            # ★ 恢复 token 使用统计
-            saved_token_stats = cache_data.get('token_stats', {
-                'input_tokens': 0, 'output_tokens': 0,
-                'reasoning_tokens': 0,
-                'cache_read': 0, 'cache_write': 0,
-                'total_tokens': 0, 'requests': 0,
-                'estimated_cost': 0.0,
-            })
-            
-            if silent and not self._conversation_history:
-                # 静默恢复：当前会话为空时直接加载到当前标签
-                self._conversation_history = history
-                self._context_summary = context_summary
-                self._session_id = cached_session_id
-                self._token_stats = saved_token_stats
-                # 恢复 todo 数据
-                if todo_data and hasattr(self, 'todo_list') and self.todo_list:
-                    self.todo_list.restore_todos(todo_data)
-                    self._ensure_todo_in_chat(self.todo_list, self.chat_layout)
-                # 更新 sessions 字典
-                if self._session_id in self._sessions:
-                    self._sessions[self._session_id]['conversation_history'] = self._conversation_history
-                    self._sessions[self._session_id]['context_summary'] = self._context_summary
-                    self._sessions[self._session_id]['token_stats'] = saved_token_stats
-                elif self._sessions:
-                    # 旧 session_id 已经变了，需要重新映射
-                    old_id = list(self._sessions.keys())[0]
-                    sdata = self._sessions.pop(old_id)
-                    sdata['conversation_history'] = self._conversation_history
-                    sdata['context_summary'] = self._context_summary
-                    sdata['token_stats'] = saved_token_stats
-                    self._sessions[self._session_id] = sdata
-                    # 更新标签数据
-                    for i in range(self.session_tabs.count()):
-                        if self.session_tabs.tabData(i) == old_id:
-                            self.session_tabs.setTabData(i, self._session_id)
-                            break
-                self._render_conversation_history()
-                self._update_token_stats_display()
-                self._update_context_stats()
-                # 自动重命名标签
-                if history:
-                    for msg in history:
-                        if msg.get('role') == 'user' and msg.get('content'):
-                            self._auto_rename_tab(msg['content'])
-                            break
-                print(f"[Workspace] 自动恢复上下文: {len(self._conversation_history)} 条消息")
-                return True
-            
-            # 非静默或当前会话非空：在新标签页中打开
-            self._save_current_session_state()
-            
-            # 创建新标签
-            self._session_counter += 1
-            scroll_area, chat_container, chat_layout = self._create_session_widgets()
-            self.session_stack.addWidget(scroll_area)
-            
-            # 用缓存文件名或首条用户消息作为标签名
-            label = f"Chat {self._session_counter}"
-            for msg in history:
-                if msg.get('role') == 'user' and msg.get('content'):
-                    short = msg['content'][:18].replace('\n', ' ').strip()
-                    if len(msg['content']) > 18:
-                        short += "..."
-                    label = short
-                    break
-            
-            tab_index = self.session_tabs.addTab(label)
-            self.session_tabs.setTabData(tab_index, cached_session_id)
-            
-            todo = self._create_todo_list(chat_container)
-            if todo_data:
-                todo.restore_todos(todo_data)
-                self._ensure_todo_in_chat(todo, chat_layout)
-            
-            self._sessions[cached_session_id] = {
-                'scroll_area': scroll_area,
-                'chat_container': chat_container,
-                'chat_layout': chat_layout,
-                'todo_list': todo,
-                'conversation_history': history,
-                'context_summary': context_summary,
-                'current_response': None,
-                'token_stats': saved_token_stats,
-            }
-            
-            # 切换到新标签
-            self._session_id = cached_session_id
-            self._conversation_history = history
-            self._context_summary = context_summary
-            self._current_response = None
-            self._token_stats = saved_token_stats
-            self.scroll_area = scroll_area
-            self.chat_container = chat_container
-            self.chat_layout = chat_layout
-            self.todo_list = todo
-            
-            self.session_tabs.blockSignals(True)
-            self.session_tabs.setCurrentIndex(tab_index)
-            self.session_tabs.blockSignals(False)
-            self.session_stack.setCurrentWidget(scroll_area)
-            
-            self._render_conversation_history()
-            self._update_token_stats_display()
-            self._update_context_stats()
-            
-            if not silent:
-                self._addStatus.emit(f"缓存已加载: {cache_file.name}")
-            
-            return True
-            
-        except Exception as e:
-            if not silent:
-                QtWidgets.QMessageBox.warning(self, "错误", f"加载缓存失败: {str(e)}")
-            else:
-                print(f"[Workspace] 加载缓存失败: {str(e)}")
-            return False
-    
-    def _load_cache_silent(self, cache_file: Path) -> bool:
-        """静默加载缓存（用于工作区自动恢复）"""
-        return self._load_cache(cache_file, silent=True)
-    
-    def _load_cache_dialog(self):
-        """显示加载缓存对话框"""
-        cache_files = sorted(
-            set(self._cache_dir.glob("session_*.json"))
-            | set(self._cache_dir.glob("archive_*.json"))
-            | set(self._cache_dir.glob("cache_*.json")),
-            key=lambda p: p.stat().st_mtime, reverse=True
-        )
-        
-        if not cache_files:
-            QtWidgets.QMessageBox.information(self, "提示", "没有找到缓存文件")
-            return
-        
-        # 创建选择对话框
-        dialog = QtWidgets.QDialog(self)
-        dialog.setWindowTitle("选择缓存文件")
-        dialog.setMinimumWidth(500)
-        
-        layout = QtWidgets.QVBoxLayout(dialog)
-        
-        # 文件列表
-        list_widget = QtWidgets.QListWidget()
-        for cache_file in cache_files:
-            # 读取文件信息
-            try:
-                with open(cache_file, 'r', encoding='utf-8') as f:
-                    data = json.load(f)
-                    msg_count = len(data.get('conversation_history', []))
-                    estimated_tokens = data.get('estimated_tokens', 0)
-                    created_at = data.get('created_at', '')
-                    if created_at:
-                        try:
-                            dt = datetime.fromisoformat(created_at)
-                            created_at = dt.strftime("%Y-%m-%d %H:%M:%S")
-                        except:
-                            pass
-                    token_info = f" | ~{estimated_tokens:,} tokens" if estimated_tokens else ""
-                    item_text = f"{cache_file.name}\n  {msg_count} 条消息{token_info} | {created_at}"
-            except:
-                item_text = cache_file.name
-            
-            item = QtWidgets.QListWidgetItem(item_text)
-            item.setData(QtCore.Qt.UserRole, cache_file)
-            list_widget.addItem(item)
-        
-        layout.addWidget(QtWidgets.QLabel("选择要加载的缓存文件:"))
-        layout.addWidget(list_widget)
-        
-        # 按钮
-        btn_layout = QtWidgets.QHBoxLayout()
-        btn_load = QtWidgets.QPushButton("加载")
-        btn_cancel = QtWidgets.QPushButton("取消")
-        btn_layout.addWidget(btn_load)
-        btn_layout.addWidget(btn_cancel)
-        layout.addLayout(btn_layout)
-        
-        def on_load():
-            current = list_widget.currentItem()
-            if current:
-                cache_file = current.data(QtCore.Qt.UserRole)
-                if self._load_cache(cache_file):
-                    dialog.accept()
-        
-        btn_load.clicked.connect(on_load)
-        btn_cancel.clicked.connect(dialog.reject)
-        
-        dialog.exec_()
-    
-    def _list_caches(self):
-        """列出所有缓存文件"""
-        cache_files = sorted(
-            set(self._cache_dir.glob("session_*.json"))
-            | set(self._cache_dir.glob("archive_*.json"))
-            | set(self._cache_dir.glob("cache_*.json")),
-            key=lambda p: p.stat().st_mtime, reverse=True
-        )
-        
-        if not cache_files:
-            QtWidgets.QMessageBox.information(self, "提示", "没有找到缓存文件")
-            return
-        
-        # 创建信息对话框
-        dialog = QtWidgets.QDialog(self)
-        dialog.setWindowTitle("缓存文件列表")
-        dialog.setMinimumSize(600, 400)
-        
-        layout = QtWidgets.QVBoxLayout(dialog)
-        
-        # 文本显示
-        text_edit = QtWidgets.QTextEdit()
-        text_edit.setReadOnly(True)
-        
-        lines = ["缓存文件列表:\n"]
-        for cache_file in cache_files:
-            try:
-                with open(cache_file, 'r', encoding='utf-8') as f:
-                    data = json.load(f)
-                    msg_count = len(data.get('conversation_history', []))
-                    created_at = data.get('created_at', '')
-                    session_id = data.get('session_id', '')
-                    estimated_tokens = data.get('estimated_tokens', 0)
-                    
-                    if created_at:
-                        try:
-                            dt = datetime.fromisoformat(created_at)
-                            created_at = dt.strftime("%Y-%m-%d %H:%M:%S")
-                        except:
-                            pass
-                    
-                    size_kb = cache_file.stat().st_size / 1024
-                    lines.append(f"  {cache_file.name}")
-                    lines.append(f"   会话ID: {session_id}")
-                    lines.append(f"   消息数: {msg_count}")
-                    if estimated_tokens:
-                        lines.append(f"   估算Token: ~{estimated_tokens:,}")
-                    lines.append(f"   创建时间: {created_at}")
-                    lines.append(f"   文件大小: {size_kb:.1f} KB")
-                    lines.append("")
-            except Exception as e:
-                lines.append(f"[err] {cache_file.name} (读取失败: {str(e)})")
-                lines.append("")
-        
-        text_edit.setPlainText("\n".join(lines))
-        layout.addWidget(text_edit)
-        
-        btn_close = QtWidgets.QPushButton("关闭")
-        btn_close.clicked.connect(dialog.accept)
-        layout.addWidget(btn_close)
-        
-        dialog.exec_()
-    
-    def _compress_to_summary(self):
-        """将旧对话压缩为摘要，减少 token 消耗"""
-        if len(self._conversation_history) <= 4:
-            QtWidgets.QMessageBox.information(self, "提示", "对话历史太短，无需压缩")
-            return
-        
-        # 确认操作
-        reply = QtWidgets.QMessageBox.question(
-            self, "确认压缩",
-            f"将把前 {len(self._conversation_history) - 4} 条对话压缩为摘要，"
-            f"保留最近 4 条完整对话。\n\n"
-            f"这样可以大幅减少 token 消耗。是否继续？",
-            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No
-        )
-        
-        if reply != QtWidgets.QMessageBox.Yes:
-            return
-        
-        # 执行压缩
-        old_messages = self._conversation_history[:-4]
-        recent_messages = self._conversation_history[-4:]
-        
-        # 生成详细摘要
-        summary_parts = ["[历史对话摘要 - 已压缩以节省 token]"]
-        
-        user_requests = []
-        ai_results = []
-        
-        for msg in old_messages:
-            role = msg.get('role', '')
-            content = msg.get('content', '')
-            
-            if role == 'user':
-                # 提取用户请求的核心（前200字符）
-                user_request = content[:200].replace('\n', ' ')
-                if len(content) > 200:
-                    user_request += "..."
-                user_requests.append(user_request)
-            
-            elif role == 'assistant' and content:
-                # 提取 AI 回复的关键信息
-                lines = [l.strip() for l in content.split('\n') if l.strip()]
-                if lines:
-                    # 取最后一行或前150字符
-                    result_summary = lines[-1][:150].replace('\n', ' ')
-                    if len(lines[-1]) > 150:
-                        result_summary += "..."
-                    ai_results.append(result_summary)
-        
-        # 合并摘要
-        if user_requests:
-            summary_parts.append(f"\n用户请求 ({len(user_requests)} 条):")
-            for i, req in enumerate(user_requests[:10], 1):  # 最多显示10条
-                summary_parts.append(f"  {i}. {req}")
-            if len(user_requests) > 10:
-                summary_parts.append(f"  ... 还有 {len(user_requests) - 10} 条请求")
-        
-        if ai_results:
-            summary_parts.append(f"\nAI 完成的任务 ({len(ai_results)} 条):")
-            for i, res in enumerate(ai_results[:10], 1):  # 最多显示10条
-                summary_parts.append(f"  {i}. {res}")
-            if len(ai_results) > 10:
-                summary_parts.append(f"  ... 还有 {len(ai_results) - 10} 条结果")
-        
-        summary_text = "\n".join(summary_parts)
-        
-        # 更新历史：用摘要替换旧对话
-        self._conversation_history = [
-            {'role': 'system', 'content': summary_text}
-        ] + recent_messages
-        
-        # 更新上下文摘要
-        self._context_summary = summary_text
-        
-        # 重新渲染
-        self._render_conversation_history()
-        
-        # 更新统计
-        self._update_context_stats()
-        
-        # 计算节省的 token
-        old_tokens = sum(self._estimate_tokens(json.dumps(msg)) for msg in old_messages)
-        new_tokens = self._estimate_tokens(summary_text)
-        saved_tokens = old_tokens - new_tokens
-        
-        QtWidgets.QMessageBox.information(
-            self, "压缩完成",
-            f"对话已压缩！\n\n"
-            f"原始: ~{old_tokens} tokens\n"
-            f"压缩后: ~{new_tokens} tokens\n"
-            f"节省: ~{saved_tokens} tokens ({saved_tokens/old_tokens*100:.1f}%)"
-        )
-    
-    # ---------- 历史渲染辅助 ----------
-    _CONTEXT_HEADERS = ('[Network structure]', '[Selected nodes]',
-                        '[网络结构]', '[选中节点]')
-
-    # ★ 分批渲染常量（借鉴 markstream-vue 的批次策略）
-    _BATCH_INITIAL = 30      # 首批渲染最后 N 条消息（用户最近看到的）
-    _BATCH_SIZE = 15          # 后续每批渲染 N 条
-    _BATCH_BUDGET_MS = 8      # 每批时间预算（毫秒）
-
-    def _render_conversation_history(self):
-        """重新渲染对话历史到 UI
-
-        ★ 分批渲染策略（借鉴 markstream-vue）：
-        1. 首批渲染最后 _BATCH_INITIAL 条消息（用户最近看到的）
-        2. 用 QTimer.singleShot(0) 模拟 idle callback，逐批渲染剩余
-        3. 每批设时间预算，超出则暂停让出主线程
-
-        处理三种数据格式：
-        1. role="user" 中嵌入 [Network structure] / [Selected nodes] 等上下文
-           → 用户文字正常显示，上下文数据放入可折叠区域
-        2. role="assistant" 以 [工具执行结果] 开头
-           → 解析每一条 [ok]/[err]/✅/❌ 行，创建折叠式 ToolCallItem
-        3. role="tool"（旧缓存格式）
-           → 先 add_tool_call 再 set_tool_result（折叠式）
-        """
-        # 清空当前显示（保留末尾 stretch）
-        while self.chat_layout.count() > 1:
-            item = self.chat_layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-
-        # 取消之前的分批渲染定时器
-        if hasattr(self, '_batch_render_timer') and self._batch_render_timer is not None:
-            self._batch_render_timer.stop()
-            self._batch_render_timer = None
-
-        messages = self._conversation_history
-        if not messages:
-            return
-
-        # ★ 预扫描：将消息分组为逻辑"轮次"（每轮 = 一组相关消息）
-        groups = self._group_messages_into_turns(messages)
-        total_groups = len(groups)
-
-        if total_groups <= self._BATCH_INITIAL:
-            # 消息量小，一次性渲染
-            self._render_message_groups(groups, 0, total_groups)
-        else:
-            # ★ 分批渲染：先渲染最后 _BATCH_INITIAL 组（用户最近看到的）
-            # 早期消息用占位符
-            early_count = total_groups - self._BATCH_INITIAL
-
-            # 插入占位符
-            self._batch_placeholder = QtWidgets.QLabel(
-                f"⏳ 加载历史消息 ({early_count} 轮)..."
-            )
-            self._batch_placeholder.setObjectName("batchPlaceholder")
-            self._batch_placeholder.setStyleSheet(
-                "color: #64748b; padding: 8px 12px; font-size: 12px; "
-                "font-style: italic; background: transparent;"
-            )
-            self._batch_placeholder.setAlignment(QtCore.Qt.AlignCenter)
-            # 插入到 stretch 之前
-            self.chat_layout.insertWidget(self.chat_layout.count() - 1,
-                                         self._batch_placeholder)
-
-            # 渲染最后 _BATCH_INITIAL 组
-            self._render_message_groups(groups, early_count, total_groups)
-
-            # 用 QTimer 分批渲染早期消息
-            self._batch_groups = groups
-            self._batch_cursor = early_count  # 从 early_count 向 0 回退
-            self._batch_insert_pos = 0  # 早期消息插入到布局头部
-            self._batch_render_timer = QtCore.QTimer(self)
-            self._batch_render_timer.setSingleShot(True)
-            self._batch_render_timer.timeout.connect(self._render_next_batch)
-            self._batch_render_timer.start(0)  # 下一帧开始
-
-    def _group_messages_into_turns(self, messages: list) -> list:
-        """将消息列表分组为逻辑轮次
-        
-        返回: list of (start_idx, end_idx) 元组
-        """
-        groups: list = []
-        i = 0
-        while i < len(messages):
-            msg = messages[i]
-            role = msg.get('role', '')
-
-            if role == 'user':
-                groups.append((i, i + 1))
-                i += 1
-            elif role == 'assistant':
-                if msg.get('tool_calls'):
-                    # 收集工具交互轮次
-                    j = i + 1
-                    while j < len(messages):
-                        m = messages[j]
-                        r = m.get('role', '')
-                        if r == 'tool':
-                            j += 1
-                        elif r == 'assistant':
-                            j += 1
-                            if not m.get('tool_calls'):
-                                break
-                        else:
-                            break
-                    groups.append((i, j))
-                    i = j
-                else:
-                    # 普通 assistant + 后续 tool 消息
-                    j = i + 1
-                    while j < len(messages) and messages[j].get('role') == 'tool':
-                        j += 1
-                    groups.append((i, j))
-                    i = j
-            elif role == 'system':
-                groups.append((i, i + 1))
-                i += 1
-            else:
-                groups.append((i, i + 1))
-                i += 1
-        return groups
-
-    def _render_message_groups(self, groups: list, start: int, end: int):
-        """渲染 [start, end) 范围内的消息组"""
-        messages = self._conversation_history
-        for gi in range(start, end):
-            si, ei = groups[gi]
-            try:
-                self._render_single_group(messages, si, ei)
-            except Exception:
-                import traceback
-                traceback.print_exc()
-
-    def _render_single_group(self, messages: list, si: int, ei: int):
-        """渲染一个消息组"""
-        msg = messages[si]
-        role = msg.get('role', '')
-        raw_content = msg.get('content', '') or ''
-        if isinstance(raw_content, list):
-            content = '\n'.join(
-                part.get('text', '') for part in raw_content
-                if isinstance(part, dict) and part.get('type') == 'text'
-            )
-        else:
-            content = raw_content
-
-        if role == 'user':
-            self._render_user_history(content)
-
-        elif role == 'assistant':
-            if msg.get('tool_calls'):
-                turn_msgs = messages[si:ei]
-                self._render_native_tool_turn(turn_msgs)
-            else:
-                tool_msgs = [messages[j] for j in range(si + 1, ei)
-                             if messages[j].get('role') == 'tool']
-
-                if content.lstrip().startswith('[工具执行结果]'):
-                    self._render_tool_summary_history(content, msg)
-                else:
-                    response = self._add_ai_response()
-                    thinking = msg.get('thinking', '')
-                    if thinking:
-                        response.add_thinking(thinking)
-                        response.thinking_section.finalize()
-                    self._render_old_tool_msgs(response, tool_msgs)
-                    self._restore_shell_widgets(response, msg)
-                    response.set_content(content)
-                    response.status_label.setText("历史")
-                    response.finalize()
-                    parts = []
-                    if thinking:
-                        parts.append("思考")
-                    if tool_msgs:
-                        parts.append(f"{len(tool_msgs)}次调用")
-                    label = f"历史 | {', '.join(parts)}" if parts else "历史"
-                    response.status_label.setText(label)
-
-        elif role == 'system' and '[历史对话摘要' in content:
-            response = self._add_ai_response()
-            response.add_collapsible("历史对话摘要", content)
-            response.status_label.setText("历史摘要")
-            response.finalize()
-            response.status_label.setText("历史摘要")
-
-    def _render_next_batch(self):
-        """分批渲染回调 — 渲染下一批早期消息（从后向前，插入到布局头部）"""
-        if not hasattr(self, '_batch_groups') or not self._batch_groups:
-            return
-        if self._batch_cursor <= 0:
-            # 全部渲染完毕，移除占位符
-            self._finish_batch_render()
-            return
-
-        batch_start = max(0, self._batch_cursor - self._BATCH_SIZE)
-        batch_end = self._batch_cursor
-        start_time = time.time()
-
-        # ★ 早期消息需要插入到占位符之前（即布局的第 0 个位置开始）
-        # 我们从 batch_start 到 batch_end 按顺序渲染，每个 widget 插入到
-        # 占位符位置之前（insert_pos 递增）
-        messages = self._conversation_history
-        insert_pos = self._batch_insert_pos  # 在此位置之前插入
-        rendered_count = 0
-
-        for gi in range(batch_start, batch_end):
-            si, ei = self._batch_groups[gi]
-            try:
-                widgets_before = self.chat_layout.count()
-                self._render_single_group(messages, si, ei)
-                widgets_after = self.chat_layout.count()
-                added = widgets_after - widgets_before
-
-                # 将新添加的 widget 移动到正确位置（占位符之前）
-                if added > 0:
-                    for _ in range(added):
-                        # 取出最后添加的 widget（在 stretch 之前）
-                        from_idx = self.chat_layout.count() - 2  # -1 是 stretch, -2 是新 widget
-                        item = self.chat_layout.takeAt(from_idx)
-                        if item and item.widget():
-                            self.chat_layout.insertWidget(insert_pos, item.widget())
-                            insert_pos += 1
-                    rendered_count += added
-            except Exception:
-                import traceback
-                traceback.print_exc()
-
-            # 时间预算检查
-            elapsed_ms = (time.time() - start_time) * 1000
-            if elapsed_ms > self._BATCH_BUDGET_MS and gi < batch_end - 1:
-                self._batch_cursor = gi + 1
-                self._batch_insert_pos = insert_pos
-                remaining = gi + 1
-                if hasattr(self, '_batch_placeholder') and self._batch_placeholder:
-                    try:
-                        self._batch_placeholder.setText(
-                            f"⏳ 加载历史消息 ({remaining} 轮)..."
-                        )
-                    except RuntimeError:
-                        pass
-                self._batch_render_timer.start(0)
-                return
-
-        self._batch_cursor = batch_start
-        self._batch_insert_pos = insert_pos
-
-        if self._batch_cursor > 0:
-            if hasattr(self, '_batch_placeholder') and self._batch_placeholder:
-                try:
-                    self._batch_placeholder.setText(
-                        f"⏳ 加载历史消息 ({self._batch_cursor} 轮)..."
-                    )
-                except RuntimeError:
-                    pass
-            self._batch_render_timer.start(0)
-        else:
-            self._finish_batch_render()
-
-    def _finish_batch_render(self):
-        """完成分批渲染，清理占位符"""
-        if hasattr(self, '_batch_placeholder') and self._batch_placeholder:
-            try:
-                self._batch_placeholder.setVisible(False)
-                self._batch_placeholder.deleteLater()
-            except RuntimeError:
-                pass
-            self._batch_placeholder = None
-        self._batch_groups = None
-        self._batch_render_timer = None
-
-    # ------------------------------------------------------------------
-    def _replay_todo_from_tool_call(self, tool_name: str, arguments_str: str):
-        """从历史工具调用中恢复 todo 项（不显示在 UI 执行列表中）
-        
-        注意：todo 数据现在通过 todo_data 字段在缓存中保存/恢复，
-        此方法仅作为兼容旧缓存的后备方案。
-        """
-        try:
-            if isinstance(arguments_str, str) and arguments_str:
-                args = json.loads(arguments_str)
-            elif isinstance(arguments_str, dict):
-                args = arguments_str
-            else:
-                return
-            if tool_name == 'add_todo':
-                tid = args.get('todo_id', '')
-                text = args.get('text', '')
-                status = args.get('status', 'pending')
-                if tid and text and hasattr(self, 'todo_list') and self.todo_list:
-                    self.todo_list.add_todo(tid, text, status)
-                    self._ensure_todo_in_chat(self.todo_list, self.chat_layout)
-            elif tool_name == 'update_todo':
-                tid = args.get('todo_id', '')
-                status = args.get('status', 'done')
-                if tid and hasattr(self, 'todo_list') and self.todo_list:
-                    self.todo_list.update_todo(tid, status)
-        except Exception:
-            pass  # 解析失败忽略
-
-    # ------------------------------------------------------------------
-    def _render_native_tool_turn(self, turn_msgs: list):
-        """渲染 Cursor 风格原生工具调用轮次
-        
-        turn_msgs 格式：
-          assistant(tool_calls) → tool → [assistant(tool_calls) → tool →] ... → assistant(reply)
-        静默工具（add_todo/update_todo）不显示在执行列表中，但会恢复 todo 数据。
-        """
-        response = self._add_ai_response()
-        tool_count = 0
-        final_content = ''
-        thinking = ''
-        final_msg = {}
-        
-        for m in turn_msgs:
-            r = m.get('role', '')
-            if r == 'assistant':
-                tc_list = m.get('tool_calls', [])
-                if tc_list:
-                    # 工具调用 assistant 消息：注册每个工具调用
-                    for tc in tc_list:
-                        fn = tc.get('function', {})
-                        name = fn.get('name', 'unknown')
-                        # 静默工具：恢复 todo 但不显示在执行列表
-                        if name in self._SILENT_TOOLS:
-                            self._replay_todo_from_tool_call(name, fn.get('arguments', ''))
-                            continue
-                        response.add_status(f"[tool]{name}")
-                        tool_count += 1
-                else:
-                    # 最终回复 assistant 消息
-                    final_content = m.get('content', '') or ''
-                    thinking = m.get('thinking', '')
-                    final_msg = m
-            elif r == 'tool':
-                tc_id = m.get('tool_call_id', '')
-                t_content = m.get('content', '') or ''
-                # 从 tool_call_id 查找对应的工具名
-                t_name = self._find_tool_name_by_id(turn_msgs, tc_id) or 'tool'
-                # 静默工具的结果也不显示
-                if t_name in self._SILENT_TOOLS:
-                    continue
-                success = not t_content.lstrip().startswith('[err]') and 'error' not in t_content[:50].lower()
-                prefix = "[ok] " if success else "[err] "
-                response.add_tool_result(t_name, f"{prefix}{t_content}")
-        
-        # 恢复 thinking
-        if thinking:
-            response.add_thinking(thinking)
-            response.thinking_section.finalize()
-        
-        # 恢复 Shell 折叠面板
-        self._restore_shell_widgets(response, final_msg)
-        
-        # AI 回复内容
-        if final_content:
-            response.set_content(final_content)
-        
-        # 状态标签
-        parts = []
-        if thinking:
-            parts.append("思考")
-        if tool_count > 0:
-            parts.append(f"{tool_count}次调用")
-        label = f"历史 | {', '.join(parts)}" if parts else "历史"
-        response.status_label.setText(label)
-        response.finalize()
-        response.status_label.setText(label)
-
-    @staticmethod
-    def _find_tool_name_by_id(messages: list, tool_call_id: str) -> str:
-        """从消息列表中根据 tool_call_id 查找对应的工具名"""
-        if not tool_call_id:
-            return ''
-        for m in messages:
-            if m.get('role') == 'assistant':
-                for tc in m.get('tool_calls', []):
-                    if tc.get('id') == tool_call_id:
-                        return tc.get('function', {}).get('name', '')
-        return ''
-
-    # ------------------------------------------------------------------
-    def _render_user_history(self, content: str):
-        """渲染用户历史消息，长上下文自动折叠"""
-        # 检查是否包含 [Network structure] 等上下文注入
-        split_pos = -1
-        header_tag = ''
-        for tag in self._CONTEXT_HEADERS:
-            pos = content.find(tag)
-            if pos != -1:
-                split_pos = pos
-                header_tag = tag
-                break
-
-        if split_pos > 0 and len(content) > 300:
-            # 用户实际输入 + 上下文注入
-            user_text = content[:split_pos].strip()
-            context_data = content[split_pos:]
-            # 显示用户实际文字
-            if user_text:
-                self._add_user_message(user_text)
-            # 上下文放进折叠区域
-            resp = self._add_ai_response()
-            resp.add_collapsible(header_tag.strip('[]'), context_data)
-            resp.status_label.setText("上下文")
-            resp.finalize()
-            resp.status_label.setText("上下文")
-        elif split_pos == 0 and len(content) > 300:
-            # 纯上下文（无用户文字），整块折叠
-            resp = self._add_ai_response()
-            resp.add_collapsible(header_tag.strip('[]'), content)
-            resp.status_label.setText("上下文")
-            resp.finalize()
-            resp.status_label.setText("上下文")
-        else:
-            self._add_user_message(content)
-
-    # ------------------------------------------------------------------
-    _TOOL_LINE_PREFIXES = ('[ok] ', '[err] ', '\u2705 ', '\u274c ')
-
-    def _render_tool_summary_history(self, content: str, msg: dict = None):
-        """渲染 [工具执行结果] 格式的 assistant 消息
-
-        格式示例：
-          [工具执行结果]
-          [ok] get_network_structure: ## 网络结构: /obj
-          网络类型: obj          ← 上一条的续行
-          节点数量: 0            ← 上一条的续行
-          [ok] create_node: /obj/geo1
-        """
-        if msg is None:
-            msg = {}
-        response = self._add_ai_response()
-
-        # 先按行分组：以 [ok]/[err]/✅/❌ 开头的行开始新条目，
-        # 其他行归到前一条目的续行
-        entries = []  # [(first_line, [continuation_lines])]
-        for line in content.split('\n'):
-            stripped = line.strip()
-            if not stripped or stripped == '[工具执行结果]':
-                # 空行或标题→如果有上一条目，添加空行到续行（保留格式）
-                if entries:
-                    entries[-1][1].append('')
-                continue
-            is_new_entry = any(stripped.startswith(p) for p in self._TOOL_LINE_PREFIXES)
-            if is_new_entry:
-                entries.append((stripped, []))
-            elif entries:
-                entries[-1][1].append(stripped)
-            # else: 没有前导条目的散行，忽略
-
-        tool_count = 0
-        for first_line, cont_lines in entries:
-            t_name = 'unknown'
-            success = True
-            # 解析前缀
-            rest = first_line
-            for prefix in self._TOOL_LINE_PREFIXES:
-                if first_line.startswith(prefix):
-                    if 'err' in prefix or '\u274c' in prefix:
-                        success = False
-                    rest = first_line[len(prefix):]
-                    break
-            # 解析 tool_name: result
-            if ':' in rest:
-                parts = rest.split(':', 1)
-                t_name = parts[0].strip()
-                first_result = parts[1].strip() if len(parts) > 1 else ''
-            else:
-                first_result = rest
-
-            # 合并续行
-            all_parts = [first_result] + cont_lines
-            t_result = '\n'.join(all_parts).strip()
-
-            # 静默工具不显示在执行列表
-            if t_name in self._SILENT_TOOLS:
-                continue
-            # 注册工具 + 设置结果
-            response.add_status(f"[tool]{t_name}")
-            tool_count += 1
-            result_prefix = "[ok] " if success else "[err] "
-            response.add_tool_result(t_name, f"{result_prefix}{t_result}")
-
-        # 恢复 Shell 折叠面板
-        self._restore_shell_widgets(response, msg)
-
-        # 恢复 thinking
-        thinking = msg.get('thinking', '')
-        if thinking:
-            response.add_thinking(thinking)
-            response.thinking_section.finalize()
-
-        # 恢复正文（[工具执行结果]之后可能还有 AI 正式回复）
-        # 找到工具摘要之后的正文部分
-        text_after_tools = ''
-        parts = content.split('\n\n')
-        for idx_p, part in enumerate(parts):
-            if not part.strip().startswith('[工具执行结果]') and not any(
-                part.strip().startswith(p) for p in self._TOOL_LINE_PREFIXES
-            ):
-                # 检查是否整段都是工具结果行
-                is_tool_block = all(
-                    any(line.strip().startswith(p) for p in self._TOOL_LINE_PREFIXES)
-                    or not line.strip()
-                    or line.strip() == '[工具执行结果]'
-                    for line in part.split('\n')
-                )
-                if not is_tool_block and part.strip():
-                    text_after_tools = '\n\n'.join(parts[idx_p:])
-                    break
-        if text_after_tools:
-            response.set_content(text_after_tools)
-
-        label_parts = []
-        if thinking:
-            label_parts.append("思考")
-        label_parts.append(f"{tool_count}次调用")
-        response.status_label.setText(f"历史 | {', '.join(label_parts)}")
-        response.finalize()
-        response.status_label.setText(f"历史 | {', '.join(label_parts)}")
-
-    # ------------------------------------------------------------------
-    def _restore_shell_widgets(self, response, msg: dict):
-        """从历史消息中恢复 Python Shell / System Shell 折叠面板"""
-        # 恢复 Python Shell
-        for ps in msg.get('python_shells', []):
-            code = ps.get('code', '')
-            raw_output = ps.get('output', '')
-            error = ps.get('error', '')
-            success = ps.get('success', True)
-            # 提取执行时间（和 _on_add_python_shell 相同逻辑）
-            exec_time = 0.0
-            clean_parts = []
-            for line in raw_output.split('\n'):
-                time_match = re.match(r'^执行时间:\s*([\d.]+)s$', line.strip())
-                if time_match:
-                    exec_time = float(time_match.group(1))
-                    continue
-                if line.strip() == '输出:':
-                    continue
-                clean_parts.append(line)
-            clean_output = '\n'.join(clean_parts).strip()
-            widget = PythonShellWidget(
-                code=code, output=clean_output, error=error,
-                exec_time=exec_time, success=success, parent=response
-            )
-            response.add_shell_widget(widget)
-
-        # 恢复 System Shell
-        for ss in msg.get('system_shells', []):
-            command = ss.get('command', '')
-            raw_output = ss.get('output', '')
-            error = ss.get('error', '')
-            success = ss.get('success', True)
-            cwd = ss.get('cwd', '')
-            exec_time = 0.0
-            exit_code = 0
-            stdout_parts = []
-            for line in raw_output.split('\n'):
-                tm = re.search(r'耗时:\s*([\d.]+)s', line)
-                cm = re.search(r'退出码:\s*(\d+)', line)
-                if tm:
-                    exec_time = float(tm.group(1))
-                if cm:
-                    exit_code = int(cm.group(1))
-                if tm or cm:
-                    continue
-                if line.strip() in ('--- stdout ---', '--- stderr ---'):
-                    continue
-                stdout_parts.append(line)
-            clean_output = '\n'.join(stdout_parts).strip()
-            widget = SystemShellWidget(
-                command=command, output=clean_output, error=error,
-                exit_code=exit_code, exec_time=exec_time,
-                success=success, cwd=cwd, parent=response
-            )
-            response.add_sys_shell_widget(widget)
-
-    # ------------------------------------------------------------------
-    def _render_old_tool_msgs(self, response, tool_msgs: list):
-        """渲染旧格式 role=tool 消息到 AIResponse"""
-        for tm in tool_msgs:
-            t_name = tm.get('name', 'unknown')
-            t_content = tm.get('content', '')
-            # 解析 tool_name:result_text
-            if ':' in t_content:
-                parts = t_content.split(':', 1)
-                t_name = parts[0].strip() or t_name
-                t_result = parts[1].strip() if len(parts) > 1 else t_content
-            else:
-                t_result = t_content
-            # 静默工具不显示在执行列表
-            if t_name in self._SILENT_TOOLS:
-                continue
-            success = not t_result.startswith('[err]') and not t_result.startswith('\u274c')
-            # 先注册工具调用
-            response.add_status(f"[tool]{t_name}")
-            result_prefix = "[ok] " if success else "[err] "
-
-            response.add_tool_result(t_name, f"{result_prefix}{t_result}")
-
-    # ===== Token 优化管理 =====
-    
-    def _on_optimize_menu(self):
-        """显示 Token 优化菜单"""
-        menu = QtWidgets.QMenu(self)
-        
-        # 立即优化
-        optimize_now_action = menu.addAction("立即压缩对话")
-        optimize_now_action.triggered.connect(self._optimize_now)
-        
-        menu.addSeparator()
-        
-        # 自动优化开关
-        auto_label = "自动压缩 [on]" if self._auto_optimize else "自动压缩"
-        auto_opt_action = menu.addAction(auto_label)
-        auto_opt_action.setCheckable(True)
-        auto_opt_action.setChecked(self._auto_optimize)
-        auto_opt_action.triggered.connect(lambda: setattr(self, '_auto_optimize', not self._auto_optimize))
-        
-        menu.addSeparator()
-        
-        # 压缩策略
-        strategy_menu = menu.addMenu("压缩策略")
-        for label, strat in [
-            ("激进 (最大节省)", CompressionStrategy.AGGRESSIVE),
-            ("平衡 (推荐)", CompressionStrategy.BALANCED),
-            ("保守 (保留细节)", CompressionStrategy.CONSERVATIVE),
-        ]:
-            action = strategy_menu.addAction(label)
-            action.setCheckable(True)
-            action.setChecked(self._optimization_strategy == strat)
-            action.triggered.connect(lambda _, s=strat: setattr(self, '_optimization_strategy', s))
-        
-        # 显示菜单
-        menu.exec_(self.btn_optimize.mapToGlobal(QtCore.QPoint(0, self.btn_optimize.height())))
-    
-    def _optimize_now(self):
-        """立即优化当前对话"""
-        if len(self._conversation_history) <= 4:
-            QtWidgets.QMessageBox.information(self, "提示", "对话历史太短，无需优化")
-            return
-        
-        # 计算优化前
-        before_tokens = self._calculate_context_tokens()
-        
-        # 执行优化
-        compressed_messages, stats = self.token_optimizer.compress_messages(
-            self._conversation_history,
-            strategy=self._optimization_strategy
-        )
-        
-        if stats['saved_tokens'] > 0:
-            self._conversation_history = compressed_messages
-            self._context_summary = compressed_messages[0].get('content', '') if compressed_messages and compressed_messages[0].get('role') == 'system' else self._context_summary
-            
-            # 重新渲染
-            self._render_conversation_history()
-            
-            # 更新统计
-            self._update_context_stats()
-            
-            # 显示结果
-            saved_percent = stats.get('saved_percent', 0)
-            QtWidgets.QMessageBox.information(
-                self, "优化完成",
-                f"对话已优化！\n\n"
-                f"原始: ~{before_tokens:,} tokens\n"
-                f"优化后: ~{stats['compressed_tokens']:,} tokens\n"
-                f"节省: ~{stats['saved_tokens']:,} tokens ({saved_percent:.1f}%)\n\n"
-                f"压缩了 {stats['compressed']} 条消息，保留 {stats['kept']} 条"
-            )
-        else:
-            QtWidgets.QMessageBox.information(self, "提示", "无需优化，对话历史已经很精简")
-
-    # ============================================================
-    # 自动更新
-    # ============================================================
-
-    _updateCheckDone = QtCore.Signal(dict)   # 检查结果
-    _updateApplyDone = QtCore.Signal(dict)   # 应用结果
-    _updateProgress = QtCore.Signal(str, int)  # (stage, percent)
-
-    def _silent_update_check(self):
-        """启动时静默检查更新（不弹窗，只在有更新时高亮按钮）"""
-        try:
-            self._updateCheckDone.connect(self._on_silent_check_result, QtCore.Qt.UniqueConnection)
-        except RuntimeError:
-            pass
-        threading.Thread(target=self._bg_check_update, daemon=True).start()
-
-    @QtCore.Slot(dict)
-    def _on_silent_check_result(self, result: dict):
-        """[主线程] 静默检查结果 → 如果有更新，高亮按钮 + 显示通知横幅"""
-        # 断开静默回调，防止和手动点击冲突
-        try:
-            self._updateCheckDone.disconnect(self._on_silent_check_result)
-        except RuntimeError:
-            pass
-        
-        if result.get('has_update') and result.get('remote_version'):
-            remote_ver = result['remote_version']
-            local_ver = result.get('local_version', '?')
-            release_name = result.get('release_name', '')
-            
-            # 1) 用醒目样式标记按钮
-            self.btn_update.setText(tr('update.new_ver', remote_ver))
-            self.btn_update.setToolTip(tr('update.new_ver_tip', remote_ver))
-            self.btn_update.setProperty("state", "available")
-            self.btn_update.style().unpolish(self.btn_update)
-            self.btn_update.style().polish(self.btn_update)
-            
-            # 2) 保存检查结果，供手动点击时直接使用
-            self._cached_update_result = result
-            
-            # 3) ★ 在输入区域上方显示更新通知横幅（含更新摘要）
-            try:
-                if hasattr(self, '_update_banner') and self._update_banner:
-                    self._update_banner.setVisible(True)
-                else:
-                    release_notes = result.get('release_notes', '').strip()
-                    self._update_banner = UpdateNotificationBanner(
-                        remote_version=remote_ver,
-                        release_name=release_name,
-                        local_version=local_ver,
-                        release_notes=release_notes,
-                    )
-                    self._update_banner.updateClicked.connect(self._on_banner_update)
-                    # 插入到输入区域布局的最顶部（batch_bar 之前）
-                    input_layout = self._batch_bar.parent().layout()
-                    if input_layout:
-                        input_layout.insertWidget(0, self._update_banner)
-                    self._update_banner.setVisible(True)
-            except Exception:
-                pass  # 横幅创建失败不影响主流程
-    
-    def _on_banner_update(self):
-        """通知横幅的"立即更新"按钮被点击"""
-        # 隐藏横幅
-        if hasattr(self, '_update_banner') and self._update_banner:
-            self._update_banner.setVisible(False)
-        # 触发更新流程
-        cached = getattr(self, '_cached_update_result', None)
-        if cached and cached.get('has_update'):
-            self._on_update_check_result(cached)
-            self._cached_update_result = None
-        else:
-            self._on_check_update()
-
-    def _on_check_update(self):
-        """点击 Update 按钮 → 后台检查更新（如果有缓存结果直接使用）"""
-        # 如果启动时已检测到新版本，直接显示结果
-        cached = getattr(self, '_cached_update_result', None)
-        if cached and cached.get('has_update'):
-            self._on_update_check_result(cached)
-            self._cached_update_result = None  # 用完清除
-            return
-        
-        self.btn_update.setEnabled(False)
-        self.btn_update.setText("检查中…")
-        
-        # 连接信号（只连一次，用 UniqueConnection 防重复）
-        try:
-            self._updateCheckDone.connect(self._on_update_check_result, QtCore.Qt.UniqueConnection)
-        except RuntimeError:
-            pass
-        
-        threading.Thread(target=self._bg_check_update, daemon=True).start()
-
-    def _bg_check_update(self):
-        """[后台线程] 调用 updater.check_update"""
-        try:
-            from ..utils.updater import check_update
-            result = check_update()
-        except Exception as e:
-            result = {'has_update': False, 'error': str(e), 'local_version': '?', 'remote_version': ''}
-        self._updateCheckDone.emit(result)
-
-    @QtCore.Slot(dict)
-    def _on_update_check_result(self, result: dict):
-        """[主线程] 处理检查结果"""
-        self.btn_update.setEnabled(True)
-        self.btn_update.setText("Update")
-        self.btn_update.setProperty("state", "")  # 恢复默认样式
-        self.btn_update.style().unpolish(self.btn_update)
-        self.btn_update.style().polish(self.btn_update)
-        
-        if result.get('error'):
-            QtWidgets.QMessageBox.warning(self, "检查更新", f"检查更新失败:\n{result['error']}")
-            return
-        
-        local_ver = result.get('local_version', '?')
-        remote_ver = result.get('remote_version', '?')
-        release_name = result.get('release_name', '')
-        release_notes = result.get('release_notes', '')
-        
-        if not result.get('has_update'):
-            QtWidgets.QMessageBox.information(
-                self, "检查更新",
-                f"当前已是最新版本 ✓\n\n"
-                f"本地版本: v{local_ver}\n"
-                f"最新 Release: v{remote_ver}"
-            )
-            return
-        
-        # ---- 有新版本，弹出确认对话框 ----
-        detail = f"本地版本: v{local_ver}\n最新 Release: v{remote_ver}"
-        if release_name:
-            detail += f"\n版本名称: {release_name}"
-        if release_notes:
-            detail += f"\n更新说明: {release_notes}"
-        detail += "\n\n⚠️ 更新后插件窗口将自动重启。\n（config、cache、trainData 目录不会被覆盖）"
-        
-        reply = QtWidgets.QMessageBox.question(
-            self, "发现新版本",
-            f"发现新版本 v{remote_ver}，是否立即更新？\n\n{detail}",
-            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel,
-            QtWidgets.QMessageBox.Cancel,
-        )
-        
-        if reply == QtWidgets.QMessageBox.Yes:
-            self._start_update()
-
-    # 更新进度文案轮播（下载阶段无百分比时使用）
-    _UPDATE_FUNNY_MESSAGES = [
-        "文件正在赶来的路上…",
-        "数据正在穿越互联网…",
-        "服务器正在翻箱倒柜找数据…",
-        "正在和后端同学对齐信息…",
-        "正在同步相关服务…",
-        "进度条正在努力工作…",
-        "正在从云里把数据拽下来…",
-        "服务器正在回忆文件放在哪…",
-        "正在构建本次请求的最佳实践…",
-        "数据已经发车，很快到站…",
-    ]
-
-    def _start_update(self):
-        """开始下载并应用更新"""
-        # 创建进度对话框，初始即用第一条搞怪文案 + 不确定进度条（动效）
-        first_msg = self._UPDATE_FUNNY_MESSAGES[0]
-        self._update_progress_dlg = QtWidgets.QProgressDialog(
-            first_msg, "取消", 0, 100, self
-        )
-        self._update_progress_dlg.setWindowTitle("更新 Houdini Agent")
-        self._update_progress_dlg.setWindowModality(QtCore.Qt.WindowModal)
-        self._update_progress_dlg.setAutoClose(False)
-        self._update_progress_dlg.setAutoReset(False)
-        self._update_progress_dlg.setMinimumDuration(0)
-        self._update_progress_dlg.setValue(0)
-        # 无 Content-Length 时只动不显示百分比：用不确定进度条
-        self._update_progress_dlg.setRange(0, 0)
-        # 文案轮播定时器（在 _on_update_progress 收到 downloading 0 时启动）
-        self._update_msg_index = 0
-        self._update_msg_timer = None
-        self._update_fade_anim = None
-        # QProgressDialog / QProgressBar 样式由全局 QSS 控制
-        
-        # 连接信号
-        try:
-            self._updateProgress.connect(self._on_update_progress, QtCore.Qt.UniqueConnection)
-            self._updateApplyDone.connect(self._on_update_apply_result, QtCore.Qt.UniqueConnection)
-        except RuntimeError:
-            pass
-        
-        threading.Thread(target=self._bg_download_and_apply, daemon=True).start()
-
-    def _bg_download_and_apply(self):
-        """[后台线程] 下载并应用更新"""
-        try:
-            from ..utils.updater import download_and_apply
-            result = download_and_apply(progress_callback=self._update_progress_cb)
-        except Exception as e:
-            result = {'success': False, 'error': str(e), 'updated_files': 0}
-        self._updateApplyDone.emit(result)
-
-    def _update_progress_cb(self, stage: str, percent: int):
-        """进度回调（从后台线程调用 → 通过信号到主线程）"""
-        self._updateProgress.emit(stage, percent)
-
-    def _stop_update_msg_timer(self):
-        """停止更新文案轮播定时器"""
-        if getattr(self, '_update_msg_timer', None) is not None:
-            self._update_msg_timer.stop()
-            self._update_msg_timer.deleteLater()
-            self._update_msg_timer = None
-        if getattr(self, '_update_fade_anim', None) is not None:
-            try:
-                self._update_fade_anim.stop()
-            except Exception:
-                pass
-            self._update_fade_anim = None
-
-    def _rotate_update_message(self):
-        """轮播搞怪文案并做淡入动效"""
-        if not hasattr(self, '_update_progress_dlg') or self._update_progress_dlg is None:
-            return
-        msgs = self._UPDATE_FUNNY_MESSAGES
-        if not msgs:
-            return
-        self._update_msg_index = (self._update_msg_index + 1) % len(msgs)
-        new_text = msgs[self._update_msg_index]
-        self._update_progress_dlg.setLabelText(new_text)
-        # 淡入动效：找到对话框里的 QLabel，用 QGraphicsOpacityEffect + QPropertyAnimation
-        label = self._update_progress_dlg.findChild(QtWidgets.QLabel)
-        if label is not None:
-            effect = label.graphicsEffect()
-            if effect is None:
-                effect = QtWidgets.QGraphicsOpacityEffect(label)
-                label.setGraphicsEffect(effect)
-            effect.setOpacity(0.28)
-            if getattr(self, '_update_fade_anim', None) is not None:
-                try:
-                    self._update_fade_anim.stop()
-                except Exception:
-                    pass
-            anim = QtCore.QPropertyAnimation(effect, b"opacity")
-            anim.setDuration(380)
-            anim.setStartValue(0.28)
-            anim.setEndValue(1.0)
-            anim.start(QtCore.QAbstractAnimation.DeleteWhenStopped)
-            self._update_fade_anim = anim
-
-    @QtCore.Slot(str, int)
-    def _on_update_progress(self, stage: str, percent: int):
-        """[主线程] 更新进度条（无 Content-Length 时不确定进度条 + 搞怪文案轮播）"""
-        if not hasattr(self, '_update_progress_dlg') or self._update_progress_dlg is None:
-            return
-        
-        if stage == 'downloading':
-            if percent == 0:
-                self._update_progress_dlg.setRange(0, 0)
-                self._update_progress_dlg.setLabelText(self._UPDATE_FUNNY_MESSAGES[0])
-                self._update_msg_index = 0
-                if getattr(self, '_update_msg_timer', None) is None:
-                    self._update_msg_timer = QtCore.QTimer(self)
-                    self._update_msg_timer.timeout.connect(self._rotate_update_message)
-                    self._update_msg_timer.start(2200)
-            elif 1 <= percent <= 99:
-                self._stop_update_msg_timer()
-                self._update_progress_dlg.setRange(0, 100)
-                self._update_progress_dlg.setValue(percent)
-                self._update_progress_dlg.setLabelText(f"正在下载… {percent}%")
-            else:
-                self._stop_update_msg_timer()
-                self._update_progress_dlg.setRange(0, 100)
-                self._update_progress_dlg.setValue(100)
-                self._update_progress_dlg.setLabelText("下载完成")
-        elif stage == 'extracting':
-            self._stop_update_msg_timer()
-            self._update_progress_dlg.setRange(0, 0)
-            self._update_progress_dlg.setLabelText("正在解压…")
-            self._update_progress_dlg.setValue(0)
-        elif stage == 'applying':
-            self._update_progress_dlg.setRange(0, 0)
-            self._update_progress_dlg.setLabelText("正在更新文件…")
-        elif stage == 'done':
-            self._stop_update_msg_timer()
-            self._update_progress_dlg.setRange(0, 100)
-            self._update_progress_dlg.setValue(100)
-            self._update_progress_dlg.setLabelText("更新完成！")
-        else:
-            self._update_progress_dlg.setValue(percent)
-            self._update_progress_dlg.setLabelText(f"{stage} ({percent}%)")
-
-    @QtCore.Slot(dict)
-    def _on_update_apply_result(self, result: dict):
-        """[主线程] 更新完成后的处理"""
-        self._stop_update_msg_timer()
-        # 关闭进度条
-        if hasattr(self, '_update_progress_dlg') and self._update_progress_dlg:
-            self._update_progress_dlg.close()
-            self._update_progress_dlg = None
-        
-        if not result.get('success'):
-            QtWidgets.QMessageBox.critical(
-                self, "更新失败",
-                f"更新过程中出现错误:\n{result.get('error', '未知错误')}"
-            )
-            return
-        
-        updated = result.get('updated_files', 0)
-        
-        # 更新成功 → 提示并重启
-        reply = QtWidgets.QMessageBox.information(
-            self, "更新成功",
-            f"已成功更新 {updated} 个文件！\n\n点击 OK 立即重启插件。",
-            QtWidgets.QMessageBox.Ok,
-        )
-        
-        # 延迟重启（让对话框关闭后再执行）
-        QtCore.QTimer.singleShot(200, self._do_restart)
-
-    def _do_restart(self):
-        """执行插件重启"""
-        try:
-            # 先保存当前工作区
-            main_win = self.window()
-            if hasattr(main_win, '_save_workspace'):
-                main_win._save_workspace()
-            
-            # 关闭当前窗口
-            main_win.force_quit = True
-            main_win.close()
-            
-            # 延迟重新打开（让窗口完全关闭后再重建）
-            # 注意：使用绝对导入的函数引用，避免模块被清除后相对导入失败
-            from houdini_agent.utils.updater import restart_plugin as _restart_fn
-            QtCore.QTimer.singleShot(500, _restart_fn)
-        except Exception as e:
-            print(f"[Updater] Restart error: {e}")
-            QtWidgets.QMessageBox.warning(
-                self, "重启失败",
-                f"自动重启失败，请手动关闭并重新打开插件。\n\n错误: {e}"
-            )
